Codec.Compression.Zlib.Lens

Description

lens support for the zlib library

High-Level API

gzipped :: Iso' ByteString ByteString

Compresses a ByteString using the gzip compression format.

gzipped = compressed gzip
gzipped = gzipped' defaultParams

zlibbed :: Iso' ByteString ByteString

Compresses a ByteString using the zlib compression format.

zlibbed = compressed zlib
zlibbed = 'zlibbed\'' defaultParams

deflated :: Iso' ByteString ByteString

Compresses a ByteString using the deflate compression format.

deflated = compressed deflate
deflated = deflated' defaultParams

compressed :: Format -> Iso' ByteString ByteString

Compresses a ByteString using the given compression format.

compressed fmt = compressed' fmt defaultParams

data Format :: *

gzip :: Format

The gzip compression format.

zlib :: Format

The zlib compression format.

deflate :: Format

The deflate compression format.

Low-Level API

zlibbed' :: Params -> Iso' ByteString ByteString

Compresses a ByteString using the zlib compression format and the given advanced parameters.

zlibbed = compressed zlib
zlibbed = zlibbed' defaultParams

gzipped' :: Params -> Iso' ByteString ByteString

Compresses a ByteString using the gzip compression format and the given advanced parameters.

gzipped = compressed gzip
gzipped = gzipped' defaultParams

deflated' :: Params -> Iso' ByteString ByteString

Compresses a ByteString using the deflate compression format and the given advanced parameters.

deflated = compressed deflate
deflated = deflated' defaultParams

compressed' :: Format -> Params -> Iso' ByteString ByteString

Compresses a ByteString using the given compression format and the given advanced parameters.

data Params

The advanced parameters needed by gzipped', zlibbed', deflated', and compressed'.

Use defaultParams and the provided Lenses to construct custom Params.

defaultParams :: Params

The default advanced parameters for compression and decompression.

levelC :: Lens' Params CompressionLevel

The compression level.

methodC :: Lens' Params Method

The compression method.

windowBitsC :: Lens' Params WindowBits

The number of bits in the compression window.

windowBitsD :: Lens' Params WindowBits

The number of bits in the decompression window.

memoryLevelC :: Lens' Params MemoryLevel

The amount of memory allowed for the internal compression state.

strategyC :: Lens' Params CompressionStrategy

The compression strategy.

bufferSizeC :: Lens' Params Int

The initial buffer size during compression.

bufferSizeD :: Lens' Params Int

The initial buffer size during decompression.

dictionary :: Lens' Params (Maybe ByteString)

Just the custom (de)compression dictionary to use, or Nothing to not use a custom dictionary.

data CompressionLevel :: *

defaultCompression :: CompressionLevel

noCompression :: CompressionLevel

bestSpeed :: CompressionLevel

bestCompression :: CompressionLevel

compressionLevel :: Int -> CompressionLevel

data Method :: *

deflateMethod :: Method

data WindowBits :: *

defaultWindowBits :: WindowBits

windowBits :: Int -> WindowBits

data MemoryLevel :: *

defaultMemoryLevel :: MemoryLevel

minMemoryLevel :: MemoryLevel

maxMemoryLevel :: MemoryLevel

memoryLevel :: Int -> MemoryLevel

data CompressionStrategy :: *

defaultStrategy :: CompressionStrategy

filteredStrategy :: CompressionStrategy

huffmanOnlyStrategy :: CompressionStrategy