Control.Applicative.Backwards

Description

Making functors with an Applicative instance that performs actions in the reverse order.

NB: This module is only included in lens for backwards compatibility with transformers versions before 3.0.

Documentation

newtype Backwards f a

The same functor, but with an Applicative instance that performs actions in the reverse order.

Constructors

Backwards

Fields

forwards :: f a