Control.Applicative.Lift

Description

Adding a new kind of pure computation to an applicative functor.

NB: This module is only included in lens for backwards compatibility with transformers versions before 3.0.

Documentation

data Lift f a

Applicative functor formed by adding pure computations to a given applicative functor.

Constructors

Pure a

Other (f a)

unLift :: Applicative f => Lift f a -> f a

Projection to the other functor.

Collecting errors

type Errors e = Lift (Constant e)

An applicative functor that collects a monoid (e.g. lists) of errors. A sequence of computations fails if any of its components do, but unlike monads made with ErrorT from Control.Monad.Trans.Error, these computations continue after an error, collecting all the errors.

failure :: Monoid e => e -> Errors e a

Report an error.