Control.Exception.Lens

Description

Handling

catching :: Getting (Endo (Maybe a)) SomeException t a b -> IO r -> (a -> IO r) -> IO r

 catching :: Prism' SomeException a     -> IO r -> (a -> IO r) -> IO r
 catching :: Lens' SomeException a      -> IO r -> (a -> IO r) -> IO r
 catching :: Traversal' SomeException a -> IO r -> (a -> IO r) -> IO r
 catching :: Iso' SomeException a       -> IO r -> (a -> IO r) -> IO r
 catching :: Getter SomeException a     -> IO r -> (a -> IO r) -> IO r
 catching :: Fold SomeException a       -> IO r -> (a -> IO r) -> IO r

catching_ :: Getting (Endo (Maybe a)) SomeException t a b -> IO r -> IO r -> IO r

 catching_ :: Prism' SomeException a     -> IO r -> IO r -> IO r
 catching_ :: Lens' SomeException a      -> IO r -> IO r -> IO r
 catching_ :: Traversal' SomeException a -> IO r -> IO r -> IO r
 catching_ :: Iso' SomeException a       -> IO r -> IO r -> IO r
 catching_ :: Getter SomeException a     -> IO r -> IO r -> IO r
 catching_ :: Fold SomeException a       -> IO r -> IO r -> IO r

handling :: Getting (Endo (Maybe a)) SomeException t a b -> (a -> IO r) -> IO r -> IO r

 handling :: Prism' SomeException a     -> (a -> IO r) -> IO r -> IO r
 handling :: Lens' SomeException a      -> (a -> IO r) -> IO r -> IO r
 handling :: Traversal' SomeException a -> (a -> IO r) -> IO r -> IO r
 handling :: Iso' SomeException a       -> (a -> IO r) -> IO r -> IO r
 handling :: Fold SomeException a       -> (a -> IO r) -> IO r -> IO r
 handling :: Getter SomeException a     -> (a -> IO r) -> IO r -> IO r

handling_ :: Getting (Endo (Maybe a)) SomeException t a b -> IO r -> IO r -> IO r

>>> handling_ nonTermination (return "caught") $ throwIO NonTermination
"caught"

 handling_ :: Prism' SomeException a     -> IO r -> IO r -> IO r
 handling_ :: Lens' SomeException a      -> IO r -> IO r -> IO r
 handling_ :: Traversal' SomeException a -> IO r -> IO r -> IO r
 handling_ :: Iso' SomeException a       -> IO r -> IO r -> IO r
 handling_ :: Getter SomeException a     -> IO r -> IO r -> IO r
 handling_ :: Fold SomeException a       -> IO r -> IO r -> IO r

Throwing

throwing :: Reviewing' SomeException t -> t -> a

Throw an Exception described by a Prism. Exceptions may be thrown from purely functional code, but may only be caught within the IO monad.

throwing l ≡ reviews l throw

>>> throwing _nonTermination ()
*** Exception: <<loop>>

 throwing :: Prism' SomeException t -> t -> a
 throwing :: Iso' SomeException t   -> t -> a

throwingIO :: Reviewing' SomeException t -> t -> IO a

A variant of throwing that can only be used within the IO monad.

throwingIO l ≡ reviews l throwIO

Although throwingIO has a type that is a specialization of the type of throwing, the two functions are subtly different:

 throwing l e seq x   ≡ throwing e
 throwingIO l e seq x ≡ x

The first example will cause the exception e to be raised, whereas the second one won't. In fact, throwingIO will only cause an exception to be raised when it is used within the IO monad. The throwingIO variant should be used in preference to throwing to raise an exception within the IO monad because it guarantees ordering with respect to other IO operations, whereas throwing does not.

 throwingIO :: Prism' SomeException t -> t -> IO a
 throwingIO :: Iso' SomeException t   -> t -> IO a

throwingTo :: ThreadId -> Reviewing' SomeException t -> t -> IO ()

throwingTo raises an exception specified by a Prism in the target thread

throwingTo thread l ≡ reviews l (throwTo thread)

 throwingTo :: ThreadId -> Prism' SomeException t -> t -> IO a
 throwingTo :: ThreadId -> Iso' SomeException t   -> t -> IO a

Exceptions

exception :: Exception a => Prism' SomeException a

Traverse the strongly typed Exception contained in SomeException where the type of your function matches the desired Exception.

 exception :: (Applicative f, Exception a, Exception b)
           => (a -> f b) -> SomeException -> f SomeException

IOExceptions

class AsIOException p f t where

Methods

ioErr :: Overloaded' p f t IOException

Unfortunately ioException is taken by base for throwing IOExceptions.

 ioErr :: Equality' IOException IOException
 ioErr :: Prism' SomeException IOException

Arithmetic Exceptions

class AsArithException p f t where

Arithmetic exceptions.

Methods

arithException :: Overloaded' p f t ArithException

 arithException :: Equality' ArithException ArithException
 arithException :: Prism'    SomeException ArithException

overflow :: AsArithException (Market' ArithException) Mutator t => Prism' t ()

overflow ≡ arithException . overflow

 overflow :: Prism' SomeException ArithException
 overflow :: Prism' SomeException ArithException

underflow :: AsArithException (Market' ArithException) Mutator t => Prism' t ()

underflow ≡ arithException . underflow

 underflow :: Prism' SomeException ArithException
 underflow :: Prism' SomeException ArithException

lossOfPrecision :: AsArithException (Market' ArithException) Mutator t => Prism' t ()

lossOfPrecision ≡ arithException . lossOfPrecision

 lossOfPrecision :: Prism' ArithException ArithException
 lossOfPrecision :: Prism' SomeException ArithException

divideByZero :: AsArithException (Market' ArithException) Mutator t => Prism' t ()

divideByZero ≡ arithException . divideByZero

 divideByZero :: Prism' ArithException ArithException
 divideByZero :: Prism' SomeException ArithException

denormal :: AsArithException (Market' ArithException) Mutator t => Prism' t ()

denormal ≡ arithException . denormal

 denormal :: Prism' ArithException ArithException
 denormal :: Prism' SomeException ArithException

Array Exceptions

class AsArrayException p f t where

Exceptions generated by array operations

Methods

arrayException :: Overloaded' p f t ArrayException

 arrayException :: Equality' ArrayException ArrayException
 arrayException :: Prism'    SomeException  ArrayException

indexOutOfBounds :: AsArrayException (Market' ArrayException) Mutator t => Prism' t String

An attempt was made to index an array outside its declared bounds.

indexOutOfBounds ≡ arrayException . indexOutOfBounds

undefinedElement :: AsArrayException (Market' ArrayException) Mutator t => Prism' t String

An attempt was made to evaluate an element of an array that had not been initialized.

undefinedElement ≡ arrayException . undefinedElement

Assertion Failed

class AsAssertionFailed p f t where

assert was applied to False.

Methods

assertionFailed :: Overloaded' p f t AssertionFailed

 assertionFailed :: Equality' AssertionFailed AssertionFailed
 assertionFailed :: Prism'    SomeException   AssertionFailed

_assertionFailed :: (AsAssertionFailed p f t, Profunctor p, Functor f) => Overloaded' p f t String

Retrieve the text of a failed assertion

AssertionFailed is isomorphic to a String

 _assertionFailed :: Iso' AssertionFailed String
 _assertionFailed :: Prism' SomeException String

Async Exceptions

class AsAsyncException p f t where

Asynchronous exceptions.

Methods

asyncException :: Overloaded' p f t AsyncException

stackOverflow :: AsAsyncException (Market' AsyncException) Mutator t => Prism' t ()

The current thread's stack exceeded its limit. Since an exception has been raised, the thread's stack will certainly be below its limit again, but the programmer should take remedial action immediately.

 stackOverflow :: Prism' AssertionFailed String
 stackOverflow :: Prism' SomeException   String

heapOverflow :: AsAsyncException (Market' AsyncException) Mutator t => Prism' t ()

The program's heap is reaching its limit, and the program should take action to reduce the amount of live data it has.

Notes:

It is undefined which thread receives this exception.

GHC currently does not throw HeapOverflow exceptions.

 heapOverflow :: Prism' AssertionFailed String
 heapOverflow :: Prism' SomeException   String

threadKilled :: AsAsyncException (Market' AsyncException) Mutator t => Prism' t ()

This exception is raised by another thread calling killThread, or by the system if it needs to terminate the thread for some reason.

 threadKilled :: Prism' AssertionFailed String
 threadKilled :: Prism' SomeException   String

userInterrupt :: AsAsyncException (Market' AsyncException) Mutator t => Prism' t ()

This exception is raised by default in the main thread of the program when the user requests to terminate the program via the usual mechanism(s) (e.g. Control-C in the console).

 userInterrupt :: Prism' AssertionFailed String
 userInterrupt :: Prism' SomeException   String

Non-Termination

class AsNonTermination p f t where

Methods

nonTermination :: Overloaded' p f t NonTermination

Thrown when the runtime system detects that the computation is guaranteed not to terminate. Note that there is no guarantee that the runtime system will notice whether any given computation is guaranteed to terminate or not.

_nonTermination :: (AsNonTermination p f t, Profunctor p, Functor f) => Overloaded' p f t ()

NonTermination is isomorphic to ()

 _nonTermination :: Iso'   NonTermination ()
 _nonTermination :: Prism' SomeException  ()

Nested Atomically

class AsNestedAtomically p f t where

Methods

nestedAtomically :: Overloaded' p f t NestedAtomically

Thrown when the program attempts to call atomically, from the stm package, inside another call to atomically.

 nestedAtomically :: Equality' SomeException NestedAtomically
 nestedAtomically :: Prism'    SomeException NestedAtomically

_nestedAtomically :: (AsNestedAtomically p f t, Profunctor p, Functor f) => Overloaded' p f t ()

NestedAtomically is isomorphic to ()

 _nestedAtomically :: Iso'   NestedAtomically ()
 _nestedAtomically :: Prism' SomeException  ()

Blocked Indefinitely

on MVar

class AsBlockedIndefinitelyOnMVar p f t where

Methods

blockedIndefinitelyOnMVar :: Overloaded' p f t BlockedIndefinitelyOnMVar

The thread is blocked on an MVar, but there are no other references to the MVar so it can't ever continue.

 blockedIndefinitelyOnMVar :: Equality' BlockedIndefinitelyOnMVar BlockedIndefinitelyOnMVar
 blockedIndefinitelyOnMVar :: Prism'    SomeException             BlockedIndefinitelyOnMVar

_blockedIndefinitelyOnMVar :: (AsBlockedIndefinitelyOnMVar p f t, Profunctor p, Functor f) => Overloaded' p f t ()

BlockedIndefinetelyOnMVar is isomorphic to ()

 _blockedIndefinitelyOnMVar :: Iso'   BlockedIndefinitelyOnMVar ()
 _blockedIndefinitelyOnMVar :: Prism' SomeException  ()

on STM

class AsBlockedIndefinitelyOnSTM p f t where

Methods

blockedIndefinitelyOnSTM :: Overloaded' p f t BlockedIndefinitelyOnSTM

The thread is waiting to retry an STM transaction, but there are no other references to any TVars involved, so it can't ever continue.

 blockedIndefinitelyOnSTM :: Equality' BlockedIndefinitelyOnSTM BlockedIndefinitelyOnSTM
 blockedIndefinitelyOnSTM :: Prism'    SomeException            BlockedIndefinitelyOnSTM

_blockedIndefinitelyOnSTM :: (AsBlockedIndefinitelyOnSTM p f t, Profunctor p, Functor f) => Overloaded' p f t ()

BlockedIndefinetelyOnSTM is isomorphic to ()

 _blockedIndefinitelyOnSTM :: Iso'   BlockedIndefinitelyOnSTM ()
 _blockedIndefinitelyOnSTM :: Prism' SomeException  ()

Deadlock

class AsDeadlock p f t where

Methods

deadlock :: Overloaded' p f t Deadlock

There are no runnable threads, so the program is deadlocked. The Deadlock exception is raised in the main thread only.

 deadlock :: Equality' Deadlock      Deadlock
 deadlock :: Prism'    SomeException Deadlock

_deadlock :: (AsDeadlock p f t, Profunctor p, Functor f) => Overloaded' p f t ()

Deadlock is isomorphic to ()

 _deadlock :: Iso'   Deadlock      ()
 _deadlock :: Prism' SomeException ()

No Such Method

class AsNoMethodError p f t where

Methods

noMethodError :: Overloaded' p f t NoMethodError

A class method without a definition (neither a default definition, nor a definition in the appropriate instance) was called.

 noMethodError :: Equality' NoMethodError NoMethodError
 noMethodError :: Prism'    SomeException NoMethodError

_noMethodError :: (AsNoMethodError p f t, Profunctor p, Functor f) => Overloaded' p f t String

Information about which method it was

NoMethodError is isomorphic to a String

 _noMethodError :: Iso'   NoMethodError String
 _noMethodError :: Prism' SomeException String

Pattern Match Failure

class AsPatternMatchFail p f t where

Methods

patternMatchFail :: Overloaded' p f t PatternMatchFail

A pattern match failed.

 patternMatchFail :: Equality' PatternMatchFail PatternMatchFail
 patternMatchFail :: Prism'    SomeException    PatternMatchFail

_patternMatchFail :: (AsPatternMatchFail p f t, Profunctor p, Functor f) => Overloaded' p f t String

Information about the source location of the pattern

PatternMatchFail is isomorphic to a String

 _patternMatchFail :: Iso'   PatternMatchFail String
 _patternMatchFail :: Prism' SomeException    String

Record

Constructor Error

class AsRecConError p f t where

Methods

recConError :: Overloaded' p f t RecConError

An uninitialised record field was used.

 recConError :: Equality' RecConError RecConError
 recConError :: Prism' SomeException RecConError

_recConError :: (AsRecConError p f t, Profunctor p, Functor f) => Overloaded' p f t String

Information about the source location where the record was constructed

RecConError is isomorphic to a String

 _recConError :: Iso'   RecConError   String
 _recConError :: Prism' SomeException String

Selection Error

class AsRecSelError p f t where

A record selector was applied to a constructor without the appropriate field. This can only happen with a datatype with multiple constructors, where some fields are in one constructor but not another.

Methods

recSelError :: Overloaded' p f t RecSelError

 recSelError :: Equality' RecSelError RecSelError
 recSelError :: Prism' SomeException RecSelError

_recSelError :: (AsRecSelError p f t, Profunctor p, Functor f) => Overloaded' p f t String

Information about the source location where the record selection occurred

RecSelError is isomorphic to a String

 _recSelError :: Iso'   RecSelError   String
 _recSelError :: Prism' SomeException String

Update Error

class AsRecUpdError p f t where

A record update was performed on a constructor without the appropriate field. This can only happen with a datatype with multiple constructors, where some fields are in one constructor but not another.

Methods

recUpdError :: Overloaded' p f t RecUpdError

 recUpdError :: Equality' RecUpdError RecUpdError
 recUpdError :: Prism' SomeException RecUpdError

_recUpdError :: (AsRecUpdError p f t, Profunctor p, Functor f) => Overloaded' p f t String

Information about the source location where the record was updated

RecUpdError is isomorphic to a String

 _recUpdError :: Iso'   RecUpdError   String
 _recUpdError :: Prism' SomeException String

Error Call

class AsErrorCall p f t where

This is thrown when the user calls error.

Methods

errorCall :: Overloaded' p f t ErrorCall

 errorCall :: Equality' ErrorCall     ErrorCall
 errorCall :: Prism'    SomeException ErrorCall

_errorCall :: (AsErrorCall p f t, Profunctor p, Functor f) => Overloaded' p f t String

Retrieve the argument given to error.

ErrorCall is isomorphic to a String

>>> catching _errorCall (error "touch down!") return
"touch down!"

 _errorCall :: Iso'   ErrorCall     String
 _errorCall :: Prism' SomeException String