Control.Lens.Each

Description

Indices

type family Index s :: *

This is a common Index type shared by Each, At, Contains and Ixed.

Each

class (Functor f, Index s ~ Index t) => Each f s t a b | s -> a, t -> b, s b -> t, t a -> s where

Extract each element of a (potentially monomorphic) container.

Notably, when applied to a tuple, this generalizes both to arbitrary homogeneous tuples.

>>> (1,2,3) & each *~ 10
(10,20,30)

It can also be used on monomorphic containers like Text or ByteString.

>>> over each Char.toUpper ("hello"^.Text.packed)
"HELLO"

each is an IndexedTraversal, so it can be used to access keys in many containers:

>>> itoListOf each $ Map.fromList [("hello",2),("world",4)]
[("hello",2),("world",4)]

>>> ("hello","world") & each.each %~ Char.toUpper
("HELLO","WORLD")

Methods

each :: IndexedLensLike (Index s) f s t a b