Control.Lens.Indexed

Description

Combinators for working with Indexed functions.

Documentation

class Profunctor c => Indexable i c where

This class permits overloading of function application for things that also admit a notion of a key or index.

Methods

indexed :: c a b -> i -> a -> b

Build a function from an Indexed function

Indexed Functions

newtype Indexed i a b

A function with access to a index. This constructor may be useful when you need to store a Indexable in a container to avoid ImpredicativeTypes.

Constructors

Indexed

Fields

unindexed :: i -> a -> b

(<.>) :: Indexable (i, j) k => (Indexed i s t -> r) -> (Indexed j a b -> s -> t) -> k a b -> r

Composition of Indexed functions

Mnemonically, the < and > points to the fact that we want to preserve the indices.

(<.) :: Indexable i k => (Indexed i s t -> r) -> ((a -> b) -> s -> t) -> k a b -> r

Compose an Indexed function with a non-indexed function.

Mnemonically, the < points to the indexing we want to preserve.

(.>) :: (st -> r) -> (kab -> st) -> kab -> r

Compose a non-indexed function with an Indexed function.

Mnemonically, the > points to the indexing we want to preserve.

Note, even if you use (.), this gives you the most recent index.

icompose :: Indexable k c => (i -> j -> k) -> (Indexed i s t -> r) -> (Indexed j a b -> s -> t) -> c a b -> r

Composition of Indexed functions with a user supplied function for combining indices

reindexed :: Indexable j k => (i -> j) -> (Indexed i a b -> r) -> k a b -> r

Remap the index.

Indexing existing lenses, traversals, etc.

indexing :: Indexable Int k => ((a -> Indexing f b) -> s -> Indexing f t) -> k a (f b) -> s -> f t

Transform a Traversal into an IndexedTraversal or a Fold into an IndexedFold, etc.

 indexing :: Traversal s t a b -> IndexedTraversal Int s t a b
 indexing :: Prism s t a b         -> IndexedTraversal Int s t a b
 indexing :: Lens s t a b           -> IndexedLens Int s t a b
 indexing :: Iso s t a b             -> IndexedLens Int s t a b
 indexing :: Fold s t               -> IndexedFold Int s t
 indexing :: Getter s t           -> IndexedGetter Int s t a b

indexing64 :: Indexable Int64 k => ((a -> Indexing64 f b) -> s -> Indexing64 f t) -> k a (f b) -> s -> f t

Transform a Traversal into an IndexedTraversal or a Fold into an IndexedFold, etc.

This combinator is like indexing except that it handles large Traversals and Folds gracefully.

 indexing64 :: Traversal s t a b -> IndexedTraversal Int64 s t a b
 indexing64 :: Prism s t a b         -> IndexedTraversal Int64 s t a b
 indexing64 :: Lens s t a b           -> IndexedLens Int64 s t a b
 indexing64 :: Iso s t a b             -> IndexedLens Int64 s t a b
 indexing64 :: Fold s t               -> IndexedFold Int64 s t
 indexing64 :: Getter s t           -> IndexedGetter Int64 s t a b

withIndex

(@~) :: (Indexed i s t -> r) -> (i -> s -> t) -> r

withIndex :: (Indexed i s t -> r) -> (i -> s -> t) -> r

type IndexedLensLike k f s t a b = k a (f b) -> s -> f t

Convenient alias for constructing indexed lenses and their ilk

type IndexedLensLike' k f s a = k a (f a) -> s -> f s

Convenient alias for constructing simple indexed lenses and their ilk