Control.Lens.IndexedAction

Description

Composable Actions

type IndexedAction i m s a = forall p f r. (Indexable i p, Effective m r f) => p a (f a) -> s -> f s

An IndexedAction is an IndexedGetter enriched with access to a Monad for side-effects.

Every Getter can be used as an Action

You can compose an Action with another Action using (.) from the Prelude.

iact :: Monad m => (s -> m (i, a)) -> IndexedAction i m s a

Construct an IndexedAction from a monadic side-effect

iperform :: Monad m => IndexedActing i m (i, a) s t a b -> s -> m (i, a)

Perform an IndexedAction.

perform ≡ flip (^@!)

iperforms :: Monad m => IndexedActing i m e s t a b -> (i -> a -> e) -> s -> m e

Perform an IndexedAction and modify the result.

(^@!) :: Monad m => s -> IndexedActing i m (i, a) s t a b -> m (i, a)

Perform an IndexedAction

>>> ["hello","world"]^!traversed.iact putStrLn
(0,hello)
(1,world)

Folds with Effects

type IndexedMonadicFold i m s a = forall p f r. (Indexable i p, Effective m r f, Applicative f) => p a (f a) -> s -> f s

An IndexedMonadicFold is an IndexedFold enriched with access to a Monad for side-effects.

Every IndexedFold can be used as an IndexedMonadicFold, that simply ignores the access to the Monad.

You can compose an IndexedMonadicFold with another IndexedMonadicFold using (.) from the Prelude.

Implementation Details

type IndexedActing i m r s t a b = Indexed i a (Effect m r b) -> s -> Effect m r t

Used to evaluate an IndexedAction.