Control.Lens.Internal

Description

These are some of the explicit Functor instances that leak into the type signatures of Control.Lens. You shouldn't need to import this module directly, unless you are coming up with a whole new kind of "Family" and need to add instances.

Documentation

newtype Focusing m s a

Used by Zoom to zoom into StateT

Constructors

Focusing

Fields

unfocusing :: m (s, a)

newtype FocusingWith w m s a

Used by Zoom to zoom into RWST

Constructors

FocusingWith

Fields

unfocusingWith :: m (s, a, w)

newtype FocusingPlus w k s a

Used by Zoom to zoom into WriterT.

Constructors

FocusingPlus

Fields

unfocusingPlus :: k (s, w) a

newtype FocusingOn f k s a

Used by Zoom to zoom into MaybeT or ListT

Constructors

FocusingOn

Fields

unfocusingOn :: k (f s) a

newtype May a

Make a monoid out of Maybe for error handling

Constructors

May

Fields

getMay :: Maybe a

newtype FocusingMay k s a

Used by Zoom to zoom into ErrorT

Constructors

FocusingMay

Fields

unfocusingMay :: k (May s) a

newtype Err e a

Make a monoid out of Either for error handling

Constructors

Err

Fields

getErr :: Either e a

newtype FocusingErr e k s a

Used by Zoom to zoom into ErrorT

Constructors

FocusingErr

Fields

unfocusingErr :: k (Err e s) a

data IndexingResult f a

The result of Indexing

Constructors

IndexingResult (f a) !Int

newtype Indexing f a

Applicative composition of State Int with a Functor, used by indexed

Constructors

Indexing

Fields

runIndexing :: Int -> IndexingResult f a

newtype Traversed f

Used internally by traverseOf_ and the like.

Constructors

Traversed

Fields

getTraversed :: f ()

newtype Sequenced m

Used internally by mapM_ and the like.

Constructors

Sequenced

Fields

getSequenced :: m ()

data Min a

Used for minimumOf

Constructors

NoMin

Min a

getMin :: Min a -> Maybe a

Obtain the minimum.

data Max a

Used for maximumOf

Constructors

NoMax

Max a

getMax :: Max a -> Maybe a

Obtain the maximum

data Context a b t

The indexed store can be used to characterize a Lens and is used by clone

Context a b t is isomorphic to newtype Context a b t = Context { runContext :: forall f. Functor f => (a -> f b) -> f t }, and to exists s. (s, Lens s t a b).

A Context is like a Lens that has already been applied to a some structure.

Constructors

Context (b -> t) a

newtype Bazaar a b t

This is used to characterize a Traversal.

a.k.a. indexed Cartesian store comonad, indexed Kleene store comonad, or an indexed FunList.

http://twanvl.nl/blog/haskell/non-regular1

Bazaar a b t is isomorphic to data Bazaar a b t = Buy t | Trade (Bazaar a b (b -> t)) a, and to exists s. (s, Traversal s t a b).

A Bazaar is like a Traversal that has already been applied to some structure.

Where a Context a b t holds an a and a function from b to t, a Bazaar a b t holds N as and a function from N bs to t.

Mnemonically, a Bazaar holds many stores and you can easily add more.

This is a final encoding of Bazaar.

Constructors

Bazaar

Fields

runBazaar :: forall f. Applicative f => (a -> f b) -> f t

bazaar :: Applicative f => (a -> f b) -> Bazaar a b t -> f t

Given an action to run for each matched pair, traverse a bazaar.

bazaar :: Traversal (Bazaar a b t) t a b

duplicateBazaar :: Bazaar a c t -> Bazaar a b (Bazaar b c t)

Bazaar is an indexed Comonad.

sell :: a -> Bazaar a b b

A trivial Bazaar.

newtype Effect m r a

Wrap a monadic effect with a phantom type argument.

Constructors

Effect

Fields

getEffect :: m r

newtype EffectRWS w st m s a

Wrap a monadic effect with a phantom type argument. Used when magnifying RWST.

Constructors

EffectRWS

Fields

getEffectRWS :: st -> m (s, st, w)

class Functor f => Gettable f where

Generalizing Const so we can apply simple Applicative transformations to it and so we can get nicer error messages

A Gettable Functor ignores its argument, which it carries solely as a phantom type parameter.

To ensure this, an instance of Gettable is required to satisfy:

id = fmap f = coerce

Methods

coerce :: f a -> f b

Replace the phantom type argument.

newtype Accessor r a

Used instead of Const to report

No instance of (Settable Accessor)

when the user attempts to misuse a Setter as a Getter, rather than a monolithic unification error.

Constructors

Accessor

Fields

runAccessor :: r

class (Monad m, Gettable f) => Effective m r f | f -> m r where

An Effective Functor ignores its argument and is isomorphic to a monad wrapped around a value.

That said, the monad is possibly rather unrelated to any Applicative structure.

Methods

effective :: Isomorphic k => k (m r) (f a)

ineffective :: Effective m r f => Isomorphic k => k (f a) (m r)

A convenient antonym that is used internally.

newtype Folding f a

A Monoid for a Gettable Applicative.

Constructors

Folding

Fields

getFolding :: f a

noEffect :: (Applicative f, Gettable f) => f a

The mempty equivalent for a Gettable Applicative Functor.

class Applicative f => Settable f where

Anything Settable must be isomorphic to the Identity Functor.

Methods

untainted :: f a -> a

untainted# :: (a -> f b) -> a -> b

tainted# :: (a -> b) -> a -> f b

newtype Mutator a

Mutator is just a renamed Identity functor to give better error messages when someone attempts to use a getter as a setter.

Most user code will never need to see this type.

Constructors

Mutator

Fields

runMutator :: a

data Level a

A basic non-empty list zipper

All combinators assume the invariant that the length stored matches the number of elements in list of items to the left, and the list of items to the left is stored reversed.

Constructors

Level !Int [a] a [a]

levelWidth :: Level a -> Int

How many entries are there in this level?

leftLevel :: Level a -> Maybe (Level a)

Pull the non-emtpy list zipper left one entry

left1Level :: Level a -> Level a

Pull the non-empty list zipper left one entry, stopping at the first entry.

leftmostLevel :: Level a -> Level a

Pull the non-empty list zipper all the way to the left.

rightmostLevel :: Level a -> Level a

Pul the non-empty list zipper all the way to the right. NB:, when given an infinite list this may not terminate.

rightLevel :: Level a -> Maybe (Level a)

Pull the non-empty list zipper right one entry.

right1Level :: Level a -> Level a

Pull the non-empty list zipper right one entry, stopping at the last entry.

focusLevel :: Functor f => (a -> f a) -> Level a -> f (Level a)

This is a Lens targeting the value that we would extract from the non-empty list zipper.

view focusLevel ≡ extract

focusLevel :: Simple Lens (Level a) a

rezipLevel :: Level a -> NonEmpty a

Zip a non-empty list zipper back up, and return the result.