Control.Lens.Internal.Action

Description

Actions

class (Monad m, Functor f, Contravariant f) => Effective m r f | f -> m r where

An Effective Functor ignores its argument and is isomorphic to a Monad wrapped around a value.

That said, the Monad is possibly rather unrelated to any Applicative structure.

Methods

effective :: m r -> f a

ineffective :: f a -> m r

newtype Effect m r a

Wrap a monadic effect with a phantom type argument.

Constructors

Effect

Fields

getEffect :: m r