Control.Lens.Internal.Bazaar

Description

Documentation

class Profunctor p => Bizarre p w | w -> p where

This class is used to run the various Bazaar variants used in this library.

Methods

bazaar :: Applicative f => p a (f b) -> w a b t -> f t

newtype Bazaar p a b t

This is used to characterize a Traversal.

a.k.a. indexed Cartesian store comonad, indexed Kleene store comonad, or an indexed FunList.

http://twanvl.nl/blog/haskell/non-regular1

A Bazaar is like a Traversal that has already been applied to some structure.

Where a Context a b t holds an a and a function from b to t, a Bazaar a b t holds N as and a function from N bs to t, (where N might be infinite).

Mnemonically, a Bazaar holds many stores and you can easily add more.

This is a final encoding of Bazaar.

Constructors

Bazaar

Fields

runBazaar :: forall f. Applicative f => p a (f b) -> f t

type Bazaar' p a = Bazaar p a a

This alias is helpful when it comes to reducing repetition in type signatures.

 type Bazaar' p a t = Bazaar p a a t

newtype BazaarT p g a b t

BazaarT is like Bazaar, except that it provides a questionable Contravariant instance To protect this instance it relies on the soundness of another Contravariant type, and usage conventions.

For example. This lets us write a suitably polymorphic and lazy taking, but there must be a better way!

Constructors

BazaarT

Fields

runBazaarT :: forall f. Applicative f => p a (f b) -> f t

type BazaarT' p g a = BazaarT p g a a

This alias is helpful when it comes to reducing repetition in type signatures.

 type BazaarT' p g a t = BazaarT p g a a t