Control.Lens.Internal.Context

Description

Documentation

class IndexedFunctor w where

Methods

ifmap :: (s -> t) -> w a b s -> w a b t

class IndexedFunctor w => IndexedComonad w where

Methods

iextract :: w a a t -> t

iduplicate :: w a c t -> w a b (w b c t)

iextend :: (w b c t -> r) -> w a c t -> w a b r

class IndexedComonad w => IndexedComonadStore w where

Methods

ipos :: w a c t -> a

ipeek :: c -> w a c t -> t

ipeeks :: (a -> c) -> w a c t -> t

iseek :: b -> w a c t -> w b c t

iseeks :: (a -> b) -> w a c t -> w b c t

iexperiment :: Functor f => (b -> f c) -> w b c t -> f t

context :: w a b t -> Context a b t

class Corepresentable p => Sellable p w where

Methods

sell :: p a (w a b b)

data Context a b t

The indexed store can be used to characterize a Lens and is used by clone.

Context a b t is isomorphic to newtype Context a b t = Context { runContext :: forall f. Functor f => (a -> f b) -> f t }, and to exists s. (s, Lens s t a b).

A Context is like a Lens that has already been applied to a some structure.

Constructors

Context (b -> t) a

type Context' a = Context a a

type Context' a s = Context a a s

newtype Pretext p a b t

Constructors

Pretext

Fields

runPretext :: forall f. Functor f => p a (f b) -> f t

type Pretext' p a = Pretext p a a

type Pretext' p a s = Pretext p a a s

newtype PretextT p g a b t

Constructors

PretextT

Fields

runPretextT :: forall f. Functor f => p a (f b) -> f t

type PretextT' p g a = PretextT p g a a

type PretextT' p g a s = PretextT p g a a s