Control.Lens.Internal.Fold

Description

Monoids for folding

newtype Folding f a

A Monoid for a Contravariant Applicative.

Constructors

Folding

Fields

getFolding :: f a

newtype Traversed a f

Used internally by traverseOf_ and the like.

The argument a of the result should not be used!

Constructors

Traversed

Fields

getTraversed :: f a

newtype TraversedF a f

Used internally by traverse1Of_ and the like.

Since: 4.16

Constructors

TraversedF

Fields

getTraversedF :: f a

newtype Sequenced a m

Used internally by mapM_ and the like.

The argument a of the result should not be used!

See 4.16 Changelog entry for the explanation of "why not Apply f =>"?

Constructors

Sequenced

Fields

getSequenced :: m a

data Leftmost a

Used for firstOf.

Constructors

LPure

LLeaf a

LStep (Leftmost a)

getLeftmost :: Leftmost a -> Maybe a

Extract the Leftmost element. This will fairly eagerly determine that it can return Just the moment it sees any element at all.

data Rightmost a

Used for lastOf.

Constructors

RPure

RLeaf a

RStep (Rightmost a)

getRightmost :: Rightmost a -> Maybe a

Extract the Rightmost element. This will fairly eagerly determine that it can return Just the moment it sees any element at all.

data ReifiedMonoid a

Constructors

ReifiedMonoid

Fields

reifiedMappend :: a -> a -> a

reifiedMempty :: a

Semigroups for folding

newtype NonEmptyDList a

Constructors

NonEmptyDList

Fields

getNonEmptyDList :: [a] -> NonEmpty a