Control.Lens.Internal.Getter

Description

Internal Classes

Getters

coerce :: (Contravariant f, Functor f) => f a -> f b

This Generalizes Const so we can apply simple Applicative transformations to it and so we can get nicer error messages.

A Functor you can coerce ignores its argument, which it carries solely as a phantom type parameter.

By the Functor and Contravariant laws, an instance of Gettable will necessarily satisfy:

id = fmap f = coerce = contramap g

noEffect :: (Contravariant f, Applicative f) => f a

The mempty equivalent for a Gettable Applicative Functor.

newtype Accessor r a

Used instead of Const to report

No instance for (Settable Accessor)

when the user attempts to misuse a Setter as a Getter, rather than a monolithic unification error.

Constructors

Accessor

Fields

runAccessor :: r