Control.Lens.Internal.Jacket

Description

This module provides internal types and functions used in the implementation of Control.Lens.Zipper. You shouldn't need to import it directly, and the exported types can be used to break Zipper invariants.

Jackets

data Jacket i t b a where

Constructors

JacketAp :: Int -> Int -> Either i y -> Either i y -> Jacket i (x -> y) b a -> Jacket i x b a -> Jacket i y b a

JacketPure :: x -> Jacket i x b a

JacketLeaf :: Int -> i -> a -> Jacket i b b a

jackl :: Jacket i x b a -> Either i x

Extract Either the Leftmost key or proof that no leaves exist by collapsing the Jacket.

jackr :: Jacket i x b a -> Either i x

Extract Either the Rightmost key or proof that no leaves exist by collapsing the Jacket.

jacket :: LensLike (Tailor Int a b) s t a b -> Iso s u (Jacket Int t b a) (Jacket j u c c)

ijacket :: Overloading (Indexed i) (->) (Tailor i a b) s t a b -> Iso s u (Jacket i t b a) (Jacket j u c c)

Tailoring

data Tailor i a b t

Constructors

Tailor Int (Int -> Jacket i t b a)

tailored :: Tailor i a b t -> Jacket i t b a

Path

data Path where

A Path into a Jacket that ends at a JacketLeaf.

Constructors

ApL :: Int -> Int -> !(Path i t y b a) -> !(Jacket i x b a) -> Path i t (x -> y) b a

ApR :: Int -> Int -> !(Jacket i (x -> y) b a) -> !(Path i t y b a) -> Path i t x b a

Start :: Path i t t b a

recompress :: Path i t b b a -> Int -> i -> a -> Jacket i t b a

Reconstruct a Jacket from a Path to a JacketLeaf and a new value for the JacketLeaf.

startl :: forall a b i r x y. Path i y x b a -> Jacket i x b a -> r -> (Path i y b b a -> Int -> i -> a -> r) -> r

startr :: forall a b i r x y. Path i y x b a -> Jacket i x b a -> r -> (Path i y b b a -> Int -> i -> a -> r) -> r

movel :: forall a b i r x y. Path i y x b a -> Jacket i x b a -> r -> (Path i y b b a -> Int -> i -> a -> r) -> r

Move left one JacketLeaf.

mover :: forall a b i r x y. Path i y x b a -> Jacket i x b a -> r -> (Path i y b b a -> Int -> i -> a -> r) -> r

Move right one JacketLeaf.