Control.Lens.Internal.Magma

Description

Magma

data Magma i t b a where

This provides a way to peek at the internal structure of a Traversal or IndexedTraversal

Constructors

MagmaAp :: Magma i (x -> y) b a -> Magma i x b a -> Magma i y b a

MagmaPure :: x -> Magma i x b a

MagmaFmap :: (x -> y) -> Magma i x b a -> Magma i y b a

Magma :: i -> a -> Magma i b b a

runMagma :: Magma i t a a -> t

Run a Magma where all the individual leaves have been converted to the expected type

Molten

newtype Molten i a b t

This is a a non-reassociating initially encoded version of Bazaar.

Constructors

Molten

Fields

runMolten :: Magma i t b a

Mafic

data Mafic a b t

This is used to generate an indexed magma from an unindexed source

By constructing it this way we avoid infinite reassociations in sums where possible.

Constructors

Mafic Int (Int -> Magma Int t b a)

runMafic :: Mafic a b t -> Magma Int t b a

Generate a Magma using from a prefix sum.

TakingWhile

data TakingWhile p g a b t

This is used to generate an indexed magma from an unindexed source

By constructing it this way we avoid infinite reassociations where possible.

Constructors

TakingWhile Bool t (Bool -> Magma () t b (Corep p a))

runTakingWhile :: Corepresentable p => TakingWhile p f a b t -> Magma () t b (Corep p a)

Generate a Magma with leaves only while the predicate holds from left to right.