Control.Lens.Internal.Reflection

Description

Reifies arbitrary terms at the type level. Based on the Functional Pearl: Implicit Configurations paper by Oleg Kiselyov and Chung-chieh Shan.

http://www.cs.rutgers.edu/~ccshan/prepose/prepose.pdf

The approach from the paper was modified to work with Data.Proxy and streamline the API by Edward Kmett and Elliott Hird.

Usage comes down to two combinators, reify and reflect.

>>> reify 6 (\p -> reflect p + reflect p)
12

The argument passed along by reify is just a data Proxy t = Proxy, so all of the information needed to reconstruct your value has been moved to the type level. This enables it to be used when constructing instances (see examples/Monoid.hs).

This version is based on the "slow" path from the reflection package, but modified to work with the same Reifies class as is provided by the "fast" path, and to make sure the parameter is Typeable.

This is necessary to work around the changes to Data.Typeable in GHC HEAD.

Documentation

class Reifies s a | s -> a where

Methods

reflect :: proxy s -> a

reifyTypeable :: Typeable a => a -> (forall s. (Typeable s, Reifies s a) => Proxy s -> r) -> r

Reify a value at the type level in a Typeable-compatible fashion, to be recovered with reflect.