Control.Lens.Internal.Review

Description

Documentation

newtype Reviewed a b

This is a profunctor used internally to implement Review

It plays a role similar to that of Accessor or Const do for Control.Lens.Getter

Constructors

Reviewed

Fields

runReviewed :: b