Control.Lens.Internal.Setter

Description

Setters

class (Applicative f, Distributive f, Traversable f) => Settable f where

Anything Settable must be isomorphic to the Identity Functor.

Methods

untainted :: f a -> a

untaintedDot :: Profunctor p => p a (f b) -> p a b

taintedDot :: Profunctor p => p a b -> p a (f b)

newtype Mutator a

Mutator is just a renamed Identity Functor to give better error messages when someone attempts to use a Getter as a Setter.

Most user code will never need to see this type.

Constructors

Mutator

Fields

runMutator :: a