Control.Lens.Tuple

Description

Tuples

class Field1 s t a b | s -> a, t -> b, s b -> t, t a -> s where

Provides access to 1st field of a tuple.

Methods

_1 :: Lens s t a b

Access the 1st field of a tuple (and possibly change its type).

>>> (1,2)^._1
1

>>> _1 .~ "hello" $ (1,2)
("hello",2)

>>> (1,2) & _1 .~ "hello"
("hello",2)

>>> _1 putStrLn ("hello","world")
hello
((),"world")

This can also be used on larger tuples as well:

>>> (1,2,3,4,5) & _1 +~ 41
(42,2,3,4,5)

_1 :: Lens (a,b) (a',b) a a'
_1 :: Lens (a,b,c) (a',b,c) a a'
_1 :: Lens (a,b,c,d) (a',b,c,d) a a'
...
_1 :: Lens (a,b,c,d,e,f,g,h,i) (a',b,c,d,e,f,g,h,i) a a'

_1 :: (Generic s, Generic t, GIxed N0 (Rep s) (Rep t) a b) => Lens s t a b

Access the 1st field of a tuple (and possibly change its type).

>>> (1,2)^._1
1

>>> _1 .~ "hello" $ (1,2)
("hello",2)

>>> (1,2) & _1 .~ "hello"
("hello",2)

>>> _1 putStrLn ("hello","world")
hello
((),"world")

This can also be used on larger tuples as well:

>>> (1,2,3,4,5) & _1 +~ 41
(42,2,3,4,5)

_1 :: Lens (a,b) (a',b) a a'
_1 :: Lens (a,b,c) (a',b,c) a a'
_1 :: Lens (a,b,c,d) (a',b,c,d) a a'
...
_1 :: Lens (a,b,c,d,e,f,g,h,i) (a',b,c,d,e,f,g,h,i) a a'

class Field2 s t a b | s -> a, t -> b, s b -> t, t a -> s where

Provides access to the 2nd field of a tuple.

Methods

_2 :: Lens s t a b

Access the 2nd field of a tuple.

>>> _2 .~ "hello" $ (1,(),3,4)
(1,"hello",3,4)

>>> (1,2,3,4) & _2 *~ 3
(1,6,3,4)

>>> _2 print (1,2)
2
(1,())

anyOf _2 :: (s -> Bool) -> (a, s) -> Bool
traverse . _2 :: (Applicative f, Traversable t) => (a -> f b) -> t (s, a) -> f (t (s, b))
foldMapOf (traverse . _2) :: (Traversable t, Monoid m) => (s -> m) -> t (b, s) -> m

_2 :: (Generic s, Generic t, GIxed N1 (Rep s) (Rep t) a b) => Lens s t a b

Access the 2nd field of a tuple.

>>> _2 .~ "hello" $ (1,(),3,4)
(1,"hello",3,4)

>>> (1,2,3,4) & _2 *~ 3
(1,6,3,4)

>>> _2 print (1,2)
2
(1,())

anyOf _2 :: (s -> Bool) -> (a, s) -> Bool
traverse . _2 :: (Applicative f, Traversable t) => (a -> f b) -> t (s, a) -> f (t (s, b))
foldMapOf (traverse . _2) :: (Traversable t, Monoid m) => (s -> m) -> t (b, s) -> m

class Field3 s t a b | s -> a, t -> b, s b -> t, t a -> s where

Provides access to the 3rd field of a tuple.

Methods

_3 :: Lens s t a b

Access the 3rd field of a tuple.

_3 :: (Generic s, Generic t, GIxed N2 (Rep s) (Rep t) a b) => Lens s t a b

Access the 3rd field of a tuple.

class Field4 s t a b | s -> a, t -> b, s b -> t, t a -> s where

Provide access to the 4th field of a tuple.

Methods

_4 :: Lens s t a b

Access the 4th field of a tuple.

_4 :: (Generic s, Generic t, GIxed N3 (Rep s) (Rep t) a b) => Lens s t a b

Access the 4th field of a tuple.

class Field5 s t a b | s -> a, t -> b, s b -> t, t a -> s where

Provides access to the 5th field of a tuple.

Methods

_5 :: Lens s t a b

Access the 5th field of a tuple.

_5 :: (Generic s, Generic t, GIxed N4 (Rep s) (Rep t) a b) => Lens s t a b

Access the 5th field of a tuple.

class Field6 s t a b | s -> a, t -> b, s b -> t, t a -> s where

Provides access to the 6th element of a tuple.

Methods

_6 :: Lens s t a b

Access the 6th field of a tuple.

_6 :: (Generic s, Generic t, GIxed N5 (Rep s) (Rep t) a b) => Lens s t a b

Access the 6th field of a tuple.

class Field7 s t a b | s -> a, t -> b, s b -> t, t a -> s where

Provide access to the 7th field of a tuple.

Methods

_7 :: Lens s t a b

Access the 7th field of a tuple.

_7 :: (Generic s, Generic t, GIxed N6 (Rep s) (Rep t) a b) => Lens s t a b

Access the 7th field of a tuple.

class Field8 s t a b | s -> a, t -> b, s b -> t, t a -> s where

Provide access to the 8th field of a tuple.

Methods

_8 :: Lens s t a b

Access the 8th field of a tuple.

_8 :: (Generic s, Generic t, GIxed N7 (Rep s) (Rep t) a b) => Lens s t a b

Access the 8th field of a tuple.

class Field9 s t a b | s -> a, t -> b, s b -> t, t a -> s where

Provides access to the 9th field of a tuple.

Methods

_9 :: Lens s t a b

Access the 9th field of a tuple.

_9 :: (Generic s, Generic t, GIxed N8 (Rep s) (Rep t) a b) => Lens s t a b

Access the 9th field of a tuple.

class Field10 s t a b | s -> a, t -> b, s b -> t, t a -> s where

Provides access to the 10th field of a tuple.

Methods

_10 :: Lens s t a b

Access the 10th field of a tuple.

_10 :: (Generic s, Generic t, GIxed N9 (Rep s) (Rep t) a b) => Lens s t a b

Access the 10th field of a tuple.

class Field11 s t a b | s -> a, t -> b, s b -> t, t a -> s where

Provides access to the 11th field of a tuple.

Methods

_11 :: Lens s t a b

Access the 11th field of a tuple.

_11 :: (Generic s, Generic t, GIxed N10 (Rep s) (Rep t) a b) => Lens s t a b

Access the 11th field of a tuple.

class Field12 s t a b | s -> a, t -> b, s b -> t, t a -> s where

Provides access to the 12th field of a tuple.

Methods

_12 :: Lens s t a b

Access the 12th field of a tuple.

_12 :: (Generic s, Generic t, GIxed N11 (Rep s) (Rep t) a b) => Lens s t a b

Access the 12th field of a tuple.

class Field13 s t a b | s -> a, t -> b, s b -> t, t a -> s where

Provides access to the 13th field of a tuple.

Methods

_13 :: Lens s t a b

Access the 13th field of a tuple.

_13 :: (Generic s, Generic t, GIxed N12 (Rep s) (Rep t) a b) => Lens s t a b

Access the 13th field of a tuple.

class Field14 s t a b | s -> a, t -> b, s b -> t, t a -> s where

Provides access to the 14th field of a tuple.

Methods

_14 :: Lens s t a b

Access the 14th field of a tuple.

_14 :: (Generic s, Generic t, GIxed N13 (Rep s) (Rep t) a b) => Lens s t a b

Access the 14th field of a tuple.

class Field15 s t a b | s -> a, t -> b, s b -> t, t a -> s where

Provides access to the 15th field of a tuple.

Methods

_15 :: Lens s t a b

Access the 15th field of a tuple.

_15 :: (Generic s, Generic t, GIxed N14 (Rep s) (Rep t) a b) => Lens s t a b

Access the 15th field of a tuple.

class Field16 s t a b | s -> a, t -> b, s b -> t, t a -> s where

Provides access to the 16th field of a tuple.

Methods

_16 :: Lens s t a b

Access the 16th field of a tuple.

_16 :: (Generic s, Generic t, GIxed N15 (Rep s) (Rep t) a b) => Lens s t a b

Access the 16th field of a tuple.

class Field17 s t a b | s -> a, t -> b, s b -> t, t a -> s where

Provides access to the 17th field of a tuple.

Methods

_17 :: Lens s t a b

Access the 17th field of a tuple.

_17 :: (Generic s, Generic t, GIxed N16 (Rep s) (Rep t) a b) => Lens s t a b

Access the 17th field of a tuple.

class Field18 s t a b | s -> a, t -> b, s b -> t, t a -> s where

Provides access to the 18th field of a tuple.

Methods

_18 :: Lens s t a b

Access the 18th field of a tuple.

_18 :: (Generic s, Generic t, GIxed N17 (Rep s) (Rep t) a b) => Lens s t a b

Access the 18th field of a tuple.

class Field19 s t a b | s -> a, t -> b, s b -> t, t a -> s where

Provides access to the 19th field of a tuple.

Methods

_19 :: Lens s t a b

Access the 19th field of a tuple.

_19 :: (Generic s, Generic t, GIxed N18 (Rep s) (Rep t) a b) => Lens s t a b

Access the 19th field of a tuple.

Strict variations

_1' :: Field1 s t a b => Lens s t a b

Strict version of _1

_2' :: Field2 s t a b => Lens s t a b

Strict version of _2

_3' :: Field3 s t a b => Lens s t a b

Strict version of _3

_4' :: Field4 s t a b => Lens s t a b

Strict version of _4

_5' :: Field5 s t a b => Lens s t a b

Strict version of _5

_6' :: Field6 s t a b => Lens s t a b

Strict version of _6

_7' :: Field7 s t a b => Lens s t a b

Strict version of _7

_8' :: Field8 s t a b => Lens s t a b

Strict version of _8

_9' :: Field9 s t a b => Lens s t a b

Strict version of _9

_10' :: Field10 s t a b => Lens s t a b

Strict version of _10

_11' :: Field11 s t a b => Lens s t a b

Strict version of _11

_12' :: Field12 s t a b => Lens s t a b

Strict version of _12

_13' :: Field13 s t a b => Lens s t a b

Strict version of _13

_14' :: Field14 s t a b => Lens s t a b

Strict version of _14

_15' :: Field15 s t a b => Lens s t a b

Strict version of _15

_16' :: Field16 s t a b => Lens s t a b

Strict version of _16

_17' :: Field17 s t a b => Lens s t a b

Strict version of _17

_18' :: Field18 s t a b => Lens s t a b

Strict version of _18

_19' :: Field19 s t a b => Lens s t a b

Strict version of _19