Control.Lens.Unsafe

Description

Documentation

class Trustworthy f

This class is only exported by this untrustworthy module, but is a superclass of Gettable.

This is required because otherwise you could construct unsafeCoerce using EvilBazaar and an illegal Gettable instance that uses undefined.