Data.ByteString.Lazy.Lens

Description

Lenses for lazy bytestrings

Documentation

packedBytes :: Simple Iso [Word8] ByteString

pack (or unpack) a list of bytes into a ByteString

pack x = x ^. packedBytes

unpack x = x ^. from packedBytes

bytes :: SimpleIndexedTraversal Int ByteString Word8

Traverse the individual bytes in a ByteString

bytes = from packedBytes . itraversed

anyOf bytes (== 0x80) :: ByteString -> Bool

packedChars :: Simple Iso String ByteString

pack (or unpack) a list of characters into a ByteString

When writing back to the ByteString it is assumed that every Char lies between '\x00' and '\xff'.

pack x = x ^. packedChars

unpack x = x ^. from packedChars

chars :: SimpleIndexedTraversal Int ByteString Char

Traverse the individual bytes in a ByteString as characters.

When writing back to the ByteString it is assumed that every Char lies between '\x00' and '\xff'.

chars = from packedChars .> itraversed

anyOf chars (== 'c') :: ByteString -> Bool