Data.ByteString.Lens

Description

Documentation

class IsByteString t where

Traversals for ByteStrings.

Methods

packedBytes :: Iso' [Word8] t

pack (or unpack) a list of bytes into a strict or lazy ByteString

pack x = x ^. packedBytes

unpack x = x ^. from packedBytes

packedChars :: Iso' String t

pack (or unpack) a list of characters into a strict or lazy ByteString

When writing back to the ByteString it is assumed that every Char lies between '\x00' and '\xff'.

pack x = x ^. packedChars

unpack x = x ^. from packedChars

bytes :: IndexedTraversal' Int t Word8

Traverse each Word8 in a strict or lazy ByteString

bytes = from packedBytes . itraversed

anyOf bytes (== 0x80) :: ByteString -> Bool

chars :: IndexedTraversal' Int t Char

Traverse the individual bytes in a strict or lazy ByteString as characters.

When writing back to the ByteString it is assumed that every Char lies between '\x00' and '\xff'.

chars = from packedChars . traverse

anyOf chars (== 'c') :: ByteString -> Bool