Data.Dynamic.Lens

Description

Documentation

dynamic :: (Typeable a, Typeable b) => Projection Dynamic Dynamic a b

Traverse the typed value contained in a Dynamic where the type required by your function matches that of the contents of the Dynamic.

>>> ()^.by dynamic
<<()>>