Data.HashMap.Lens

Description

Documentation

at :: (Eq k, Hashable k) => k -> SimpleIndexedLens k (HashMap k v) (Maybe v)

This Lens can be used to read, write or delete the value associated with a key in a HashMap.

>>> :m + Control.Lens Data.HashMap.Lens

>>> HashMap.fromList [("hello",12)] ^.at "hello"
Just 12

>>> at 10 .~ Just "hello" $ HashMap.empty
fromList [(10,"hello")]

traverseMap :: IndexedTraversal k (HashMap k v) (HashMap k v') v v'

Traversal of a HashMap indexed by the key.

traverseMap = index traverseWithKey

traverseAt :: (Eq k, Hashable k) => k -> SimpleIndexedTraversal k (HashMap k v) v

Traverse the value at a given key in a HashMap