Data.IntMap.Lens

Description

Documentation

at :: Int -> SimpleIndexedLens Int (IntMap v) (Maybe v)

This Lens can be used to read, write or delete the value associated with a key in an IntMap.

>>> fromList [(1,"hello")] ^.at 1
Just "hello"

>>> at 1 .~ Just "hello" $ IntMap.empty
fromList [(1,"hello")]

 at :: Int -> (Maybe v -> f (Maybe v)) -> IntMap v -> f (IntMap v)

traverseIntMap :: IndexedTraversal Int (IntMap v) (IntMap v') v v'

Traversal of an IntMap indexed by the key.

traverseAt :: Int -> SimpleIndexedTraversal Int (IntMap v) v

Traverse the value at a given key in an IntMap

 traverseAt :: Applicative f => Int -> (v -> f v) -> IntMap v -> f (IntMap v)
 traverseAt k = at k . traverse

traverseAtMin :: SimpleIndexedTraversal Int (IntMap v) v

Traverse the value at the minimum key in a Map

The key of the minimum element is available as the index.

traverseAtMax :: SimpleIndexedTraversal Int (IntMap v) v

Traverse the value at the maximum key in a Map