Data.Pair.Lens

Description

Lenses for working with products.

Due to their ubiquity, _1 and _2 are defined in Control.Lens.

Documentation

both :: Traversal (a, a) (b, b) a b

Traverse both parts of a tuple with matching types.

value :: (k -> Bool) -> SimpleIndexedTraversal k (k, v) v

This provides a Traversal that checks a predicate on a key before allowing you to traverse into a value.