Data.Sequence.Lens

Description

Documentation

ordinal :: Int -> SimpleIndexedLens Int (Seq a) a

A Lens that can access the nth element of a Seq.

Note: This is only a legal lens if there is already such an element!

viewL :: Iso (Seq a) (Seq b) (ViewL a) (ViewL b)

A Seq is isomorphic to a ViewL

viewl m = m ^. viewL

viewR :: Iso (Seq a) (Seq b) (ViewR a) (ViewR b)

A Seq is isomorphic to a ViewR

viewr m = m ^. viewR

traverseHead :: SimpleIndexedTraversal Int (Seq a) a

Traverse the head of a Seq

traverseTail :: SimpleIndexedTraversal Int (Seq a) a

Traverse the tail of a Seq

traverseLast :: SimpleIndexedTraversal Int (Seq a) a

Traverse the last element of a Seq

traverseInit :: SimpleIndexedTraversal Int (Seq a) a

Traverse all but the last element of a Seq

traverseTo :: Int -> SimpleIndexedTraversal Int (Seq a) a

Traverse the first n elements of a Seq

traverseFrom :: Int -> SimpleIndexedTraversal Int (Seq a) a

Traverse all but the first n elements of a Seq

traverseSlice :: Int -> Int -> SimpleIndexedTraversal Int (Seq a) a

Travere all the elements numbered from i to j of a Seq