Data.Tree.Lens

Description

Documentation

root :: Simple Lens (Tree a) a

A Lens that focuses on the root of a Tree.

>>> view root $ Node 42 []
42

branches :: SimpleIndexedTraversal Int (Tree a) (Tree a)

A Traversal of the direct descendants of the root of a Tree indexed by its position in the list of children

toListOf branches ≡ subForest