Data.Typeable.Lens

Description

Documentation

_cast :: (Typeable a, Typeable b) => Simple Traversal a b

A Simple Traversal for working with a cast of a Typeable value.

_gcast :: (Typeable a, Typeable b) => Simple Traversal (c a) (c b)

A Simple Traversal for working with a cast of a Typeable value.