Data.Vector.Lens

Description

This module provides lenses and traversals for working with generic vectors.

Documentation

toVectorOf :: Getting (Endo [a]) s a -> s -> Vector a

Similar to toListOf, but returning a Vector.

>>> toVectorOf both (8,15)
fromList [8,15]

Isomorphisms

vector :: Iso [a] [b] (Vector a) (Vector b)

Convert a list to a Vector (or back)

>>> [1,2,3] ^. vector
fromList [1,2,3]

>>> [1,2,3] ^. vector . from vector
[1,2,3]

>>> Vector.fromList [0,8,15] ^. from vector . vector
fromList [0,8,15]

forced :: Iso (Vector a) (Vector b) (Vector a) (Vector b)

Convert a Vector to a version that doesn't retain any extra memory.

Lenses

sliced

Arguments

:: Int
i starting index

-> Int
n length

-> Lens' (Vector a) (Vector a)

sliced i n provides a Lens that edits the n elements starting at index i from a Lens.

This is only a valid Lens if you do not change the length of the resulting Vector.

Attempting to return a longer or shorter vector will result in violations of the Lens laws.

>>> Vector.fromList [1..10] ^. sliced 2 5
fromList [3,4,5,6,7]

>>> Vector.fromList [1..10] & sliced 2 5 . mapped .~ 0
fromList [1,2,0,0,0,0,0,8,9,10]

Traversal of individual indices

ordinals :: [Int] -> IndexedTraversal' Int (Vector a) a

This Traversal will ignore any duplicates in the supplied list of indices.

>>> toListOf (ordinals [1,3,2,5,9,10]) $ Vector.fromList [2,4..40]
[4,8,6,12,20,22]