Language.Haskell.TH.Lens

Description

Lenses, Prisms, and Traversals for working with Template Haskell

Traversals

class HasName t where

Has a Name

Methods

name :: Lens' t Name

Extract (or modify) the Name of something

class HasTypeVars t where

Provides for the extraction of free type variables, and alpha renaming.

Methods

typeVarsEx :: Set Name -> Traversal' t Name

When performing substitution into this traversal you're not allowed to substitute in a name that is bound internally or you'll violate the Traversal laws, when in doubt generate your names with newName.

class SubstType t where

Provides substitution for types

Methods

substType :: Map Name Type -> t -> t

Perform substitution for types

typeVars :: HasTypeVars t => Traversal' t Name

Traverse free type variables

substTypeVars :: HasTypeVars t => Map Name Name -> t -> t

Substitute using a map of names in for free type variables

conFields :: Traversal' Con StrictType

Provides a Traversal of the types of each field of a constructor.

conNamedFields :: Traversal' Con VarStrictType

Traversal of the types of the named fields of a constructor.

Lenses

Loc Lenses

locFileName :: Lens' Loc String

locPackage :: Lens' Loc String

locModule :: Lens' Loc String

locStart :: Lens' Loc CharPos

locEnd :: Lens' Loc CharPos

FunDep Lenses

funDepInputs :: Lens' FunDep [Name]

funDepOutputs :: Lens' FunDep [Name]

Match Lenses

matchPattern :: Lens' Match Pat

matchBody :: Lens' Match Body

matchDeclarations :: Lens' Match [Dec]

Fixity Lenses

fixityPrecedence :: Lens' Fixity Int

fixityDirection :: Lens' Fixity FixityDirection

Clause Lenses

clausePattern :: Lens' Clause [Pat]

clauseBody :: Lens' Clause Body

clauseDecs :: Lens' Clause [Dec]

FieldExp Lenses

fieldExpName :: Lens' FieldExp Name

fieldExpExpression :: Lens' FieldExp Exp

FieldPat Lenses

fieldPatName :: Lens' FieldPat Name

fieldPatPattern :: Lens' FieldPat Pat

TySynEqn Lenses

tySynEqnPatterns :: Lens' TySynEqn [Type]

tySynEqnResult :: Lens' TySynEqn Type

Prisms

Info Prisms

_ClassI :: Prism' Info (Dec, [InstanceDec])

_ClassOpI :: Prism' Info (Name, Type, ParentName, Fixity)

_TyConI :: Prism' Info Dec

_FamilyI :: Prism' Info (Dec, [InstanceDec])

_PrimTyConI :: Prism' Info (Name, Arity, Unlifted)

_DataConI :: Prism' Info (Name, Type, ParentName, Fixity)

_VarI :: Prism' Info (Name, Type, Maybe Dec, Fixity)

_TyVarI :: Prism' Info (Name, Type)

Dec Prisms

_FunD :: Prism' Dec (Name, [Clause])

_ValD :: Prism' Dec (Pat, Body, [Dec])

_DataD :: Prism' Dec (Cxt, Name, [TyVarBndr], [Con], [Name])

_NewtypeD :: Prism' Dec (Cxt, Name, [TyVarBndr], Con, [Name])

_TySynD :: Prism' Dec (Name, [TyVarBndr], Type)

_ClassD :: Prism' Dec (Cxt, Name, [TyVarBndr], [FunDep], [Dec])

_InstanceD :: Prism' Dec (Cxt, Type, [Dec])

_SigD :: Prism' Dec (Name, Type)

_ForeignD :: Prism' Dec Foreign

_InfixD :: Prism' Dec (Fixity, Name)

_PragmaD :: Prism' Dec Pragma

_FamilyD :: Prism' Dec (FamFlavour, Name, [TyVarBndr], Maybe Kind)

_DataInstD :: Prism' Dec (Cxt, Name, [Type], [Con], [Name])

_NewtypeInstD :: Prism' Dec (Cxt, Name, [Type], Con, [Name])

_TySynInstD :: Prism' Dec (Name, TySynEqn)

_ClosedTypeFamilyD :: Prism' Dec (Name, [TyVarBndr], Maybe Kind, [TySynEqn])

_RoleAnnotD :: Prism' Dec (Name, [Role])

Con Prisms

_NormalC :: Prism' Con (Name, [StrictType])

_RecC :: Prism' Con (Name, [VarStrictType])

_InfixC :: Prism' Con (StrictType, Name, StrictType)

_ForallC :: Prism' Con ([TyVarBndr], Cxt, Con)

Strict Prisms

_IsStrict :: Prism' Strict ()

_NotStrict :: Prism' Strict ()

_Unpacked :: Prism' Strict ()

Foreign Prisms

_ImportF :: Prism' Foreign (Callconv, Safety, String, Name, Type)

_ExportF :: Prism' Foreign (Callconv, String, Name, Type)

Callconv Prisms

_CCall :: Prism' Callconv ()

_StdCall :: Prism' Callconv ()

Safety Prisms

_Unsafe :: Prism' Safety ()

_Safe :: Prism' Safety ()

_Interruptible :: Prism' Safety ()

Pragma Prisms

_InlineP :: Prism' Pragma (Name, Inline, RuleMatch, Phases)

_SpecialiseP :: Prism' Pragma (Name, Type, Maybe Inline, Phases)

_SpecialiseInstP :: Prism' Pragma Type

_RuleP :: Prism' Pragma (String, [RuleBndr], Exp, Exp, Phases)

_AnnP :: Prism' Pragma (AnnTarget, Exp)

Inline Prisms

_NoInline :: Prism' Inline ()

_Inline :: Prism' Inline ()

_Inlinable :: Prism' Inline ()

RuleMatch Prisms

_ConLike :: Prism' RuleMatch ()

_FunLike :: Prism' RuleMatch ()

Phases Prisms

_AllPhases :: Prism' Phases ()

_FromPhase :: Prism' Phases Int

_BeforePhase :: Prism' Phases Int

RuleBndr Prisms

_RuleVar :: Prism' RuleBndr Name

_TypedRuleVar :: Prism' RuleBndr (Name, Type)

AnnTarget Prisms

_ModuleAnnotation :: Prism' AnnTarget ()

_TypeAnnotation :: Prism' AnnTarget Name

_ValueAnnotation :: Prism' AnnTarget Name

FunDep Prisms TODO make a lens

_FunDep :: Prism' FunDep ([Name], [Name])

FamFlavour Prisms

_TypeFam :: Prism' FamFlavour ()

_DataFam :: Prism' FamFlavour ()

FixityDirection Prisms

_InfixL :: Prism' FixityDirection ()

_InfixR :: Prism' FixityDirection ()

_InfixN :: Prism' FixityDirection ()

Exp Prisms

_VarE :: Prism' Exp Name

_ConE :: Prism' Exp Name

_LitE :: Prism' Exp Lit

_AppE :: Prism' Exp (Exp, Exp)

_InfixE :: Prism' Exp (Maybe Exp, Exp, Maybe Exp)

_UInfixE :: Prism' Exp (Exp, Exp, Exp)

_ParensE :: Prism' Exp Exp

_LamE :: Prism' Exp ([Pat], Exp)

_LamCaseE :: Prism' Exp [Match]

_TupE :: Prism' Exp [Exp]

_UnboxedTupE :: Prism' Exp [Exp]

_CondE :: Prism' Exp (Exp, Exp, Exp)

_MultiIfE :: Prism' Exp [(Guard, Exp)]

_LetE :: Prism' Exp ([Dec], Exp)

_CaseE :: Prism' Exp (Exp, [Match])

_DoE :: Prism' Exp [Stmt]

_CompE :: Prism' Exp [Stmt]

_ArithSeqE :: Prism' Exp Range

_ListE :: Prism' Exp [Exp]

_SigE :: Prism' Exp (Exp, Type)

_RecConE :: Prism' Exp (Name, [FieldExp])

_RecUpdE :: Prism' Exp (Exp, [FieldExp])

Body Prisms

_GuardedB :: Prism' Body [(Guard, Exp)]

_NormalB :: Prism' Body Exp

Guard Prisms

_NormalG :: Prism' Guard Exp

_PatG :: Prism' Guard [Stmt]

Stmt Prisms

_BindS :: Prism' Stmt (Pat, Exp)

_LetS :: Prism' Stmt [Dec]

_NoBindS :: Prism' Stmt Exp

_ParS :: Prism' Stmt [[Stmt]]

Range Prisms

_FromR :: Prism' Range Exp

_FromThenR :: Prism' Range (Exp, Exp)

_FromToR :: Prism' Range (Exp, Exp)

_FromThenToR :: Prism' Range (Exp, Exp, Exp)

Lit Prisms

_CharL :: Prism' Lit Char

_StringL :: Prism' Lit String

_IntegerL :: Prism' Lit Integer

_RationalL :: Prism' Lit Rational

_IntPrimL :: Prism' Lit Integer

_WordPrimL :: Prism' Lit Integer

_FloatPrimL :: Prism' Lit Rational

_DoublePrimL :: Prism' Lit Rational

_StringPrimL :: Prism' Lit [Word8]

Pat Prisms

_LitP :: Prism' Pat Lit

_VarP :: Prism' Pat Name

_TupP :: Prism' Pat [Pat]

_UnboxedTupP :: Prism' Pat [Pat]

_ConP :: Prism' Pat (Name, [Pat])

_InfixP :: Prism' Pat (Pat, Name, Pat)

_UInfixP :: Prism' Pat (Pat, Name, Pat)

_ParensP :: Prism' Pat Pat

_TildeP :: Prism' Pat Pat

_BangP :: Prism' Pat Pat

_AsP :: Prism' Pat (Name, Pat)

_WildP :: Prism' Pat ()

_RecP :: Prism' Pat (Name, [FieldPat])

_ListP :: Prism' Pat [Pat]

_SigP :: Prism' Pat (Pat, Type)

_ViewP :: Prism' Pat (Exp, Pat)

Type Prisms

_ForallT :: Prism' Type ([TyVarBndr], Cxt, Type)

_AppT :: Prism' Type (Type, Type)

_SigT :: Prism' Type (Type, Kind)

_VarT :: Prism' Type Name

_ConT :: Prism' Type Name

_PromotedT :: Prism' Type Name

_TupleT :: Prism' Type Int

_UnboxedTupleT :: Prism' Type Int

_ArrowT :: Prism' Type ()

_ListT :: Prism' Type ()

_PromotedTupleT :: Prism' Type Int

_PromotedNilT :: Prism' Type ()

_PromotedConsT :: Prism' Type ()

_StarT :: Prism' Type ()

_ConstraintT :: Prism' Type ()

_LitT :: Prism' Type TyLit

TyVarBndr Prisms

_PlainTV :: Prism' TyVarBndr Name

_KindedTV :: Prism' TyVarBndr (Name, Kind)

TyLit Prisms

_NumTyLit :: Prism' TyLit Integer

_StrTyLit :: Prism' TyLit String

Pred Prisms

_ClassP :: Prism' Pred (Name, [Type])

_EqualP :: Prism' Pred (Type, Type)

Role Prisms

_NominalR :: Prism' Role ()

_RepresentationalR :: Prism' Role ()