System.FilePath.Lens

Description

Documentation

(</>~) :: Setting s t FilePath FilePath -> FilePath -> s -> t

Modify the path by adding another path.

>>> :m + Control.Lens
>>> both </>~ "!!!" $ ("hello","world")
("hello/!!!","world/!!!")

 (</>~) :: Setter s a FilePath FilePath -> FilePath -> s -> a
 (</>~) :: Iso s a FilePath FilePath -> FilePath -> s -> a
 (</>~) :: Lens s a FilePath FilePath -> FilePath -> s -> a
 (</>~) :: Traversal s a FilePath FilePath -> FilePath -> s -> a

(<</>~) :: LensLike ((,) FilePath) s a FilePath FilePath -> FilePath -> s -> (FilePath, a)

Add a path onto the end of the target of a Lens and return the result

When you do not need the result of the operation, (</>~) is more flexible.

(<.>~) :: Setting s a FilePath FilePath -> String -> s -> a

Modify the path by adding extension.

>>> :m + Control.Lens
>>> both <.>~ "!!!" $ ("hello","world")
("hello.!!!","world.!!!")

 (<.>~) :: Setter s a FilePath FilePath -> String -> s -> a
 (<.>~) :: Iso s a FilePath FilePath -> String -> s -> a
 (<.>~) :: Lens s a FilePath FilePath -> String -> s -> a
 (<.>~) :: Traversal s a FilePath FilePath -> String -> s -> a

(<<.>~) :: LensLike ((,) FilePath) s a FilePath FilePath -> String -> s -> (FilePath, a)

Add an extension onto the end of the target of a Lens and return the result

When you do not need the result of the operation, (<.>~) is more flexible.

(</>=) :: MonadState s m => SimpleSetting s FilePath -> FilePath -> m ()

Modify the target(s) of a Simple Lens, Iso, Setter or Traversal by adding a path.

 (</>=) :: MonadState s m => Simple Setter s FilePath -> FilePath -> m ()
 (</>=) :: MonadState s m => Simple Iso s FilePath -> FilePath -> m ()
 (</>=) :: MonadState s m => Simple Lens s FilePath -> FilePath -> m ()
 (</>=) :: MonadState s m => Simple Traversal s FilePath -> FilePath -> m ()

(<</>=) :: MonadState s m => SimpleLensLike ((,) FilePath) s FilePath -> FilePath -> m FilePath

Add a path onto the end of the target of a Lens into your monad's state and return the result.

When you do not need the result of the operation, (</>=) is more flexible.

(<.>=) :: MonadState s m => SimpleSetting s FilePath -> String -> m ()

Modify the target(s) of a Simple Lens, Iso, Setter or Traversal by adding an extension.

 (<.>=) :: MonadState s m => Simple Setter s FilePath -> String -> m ()
 (<.>=) :: MonadState s m => Simple Iso s FilePath -> String -> m ()
 (<.>=) :: MonadState s m => Simple Lens s FilePath -> String -> m ()
 (<.>=) :: MonadState s m => Simple Traversal s FilePath -> String -> m ()

(<<.>=) :: MonadState s m => SimpleLensLike ((,) FilePath) s FilePath -> String -> m FilePath

Add an extension onto the end of the target of a Lens into your monad's state and return the result.

When you do not need the result of the operation, (<.>=) is more flexible.

basename :: Simple Lens FilePath FilePath

A lens reading and writing to the basename.

>>> basename .~ "filename" $ "path/name.png"
"path/filename.png"

directory :: Simple Lens FilePath FilePath

A lens reading and writing to the directory.

>>> "long/path/name.txt" ^. directory
"long/path"

extension :: Simple Lens FilePath FilePath

A lens reading and writing to the extension.

>>> extension .~ ".png" $ "path/name.txt"
"path/name.png"

filename :: Simple Lens FilePath FilePath

A lens reading and writing to the full filename.

>>> filename .~ "name.txt" $ "path/name.png"
"path/name.txt"