Control.Exception.Lens

Description

Control.Exception provides an example of a large open hierarchy that we can model with prisms and isomorphisms.

Additional combinators for working with IOException results can be found in System.IO.Error.Lens.

Handling

catching :: Getting (Endo (Maybe a)) SomeException t a b -> IO r -> (a -> IO r) -> IO r

Catch exceptions that match a given Prism (or any Getter, really).

>>> catching _AssertionFailed (assert False (return "uncaught")) $ \ _ -> return "caught"
"caught"

 catching :: Prism' SomeException a     -> IO r -> (a -> IO r) -> IO r
 catching :: Lens' SomeException a      -> IO r -> (a -> IO r) -> IO r
 catching :: Traversal' SomeException a -> IO r -> (a -> IO r) -> IO r
 catching :: Iso' SomeException a       -> IO r -> (a -> IO r) -> IO r
 catching :: Getter SomeException a     -> IO r -> (a -> IO r) -> IO r
 catching :: Fold SomeException a       -> IO r -> (a -> IO r) -> IO r

catching_ :: Getting (Endo (Maybe a)) SomeException t a b -> IO r -> IO r -> IO r

Catch exceptions that match a given Prism (or any Getter), discarding the information about the match. This is particuarly useful when you have a Prism' SomeException () where the result of the prism or fold isn't particularly valuable, just the fact that it matches.

>>> catching_ _AssertionFailed (assert False (return "uncaught")) $ return "caught"
"caught"

 catching_ :: Prism' SomeException a     -> IO r -> IO r -> IO r
 catching_ :: Lens' SomeException a      -> IO r -> IO r -> IO r
 catching_ :: Traversal' SomeException a -> IO r -> IO r -> IO r
 catching_ :: Iso' SomeException a       -> IO r -> IO r -> IO r
 catching_ :: Getter SomeException a     -> IO r -> IO r -> IO r
 catching_ :: Fold SomeException a       -> IO r -> IO r -> IO r

handling :: Getting (Endo (Maybe a)) SomeException t a b -> (a -> IO r) -> IO r -> IO r

A version of catching with the arguments swapped around; useful in situations where the code for the handler is shorter.

>>> handling _NonTermination (\_ -> return "caught") $ throwIO NonTermination
"caught"

 handling :: Prism' SomeException a     -> (a -> IO r) -> IO r -> IO r
 handling :: Lens' SomeException a      -> (a -> IO r) -> IO r -> IO r
 handling :: Traversal' SomeException a -> (a -> IO r) -> IO r -> IO r
 handling :: Iso' SomeException a       -> (a -> IO r) -> IO r -> IO r
 handling :: Fold SomeException a       -> (a -> IO r) -> IO r -> IO r
 handling :: Getter SomeException a     -> (a -> IO r) -> IO r -> IO r

handling_ :: Getting (Endo (Maybe a)) SomeException t a b -> IO r -> IO r -> IO r

A version of catching_ with the arguments swapped around; useful in situations where the code for the handler is shorter.

>>> handling_ _NonTermination (return "caught") $ throwIO NonTermination
"caught"

 handling_ :: Prism' SomeException a     -> IO r -> IO r -> IO r
 handling_ :: Lens' SomeException a      -> IO r -> IO r -> IO r
 handling_ :: Traversal' SomeException a -> IO r -> IO r -> IO r
 handling_ :: Iso' SomeException a       -> IO r -> IO r -> IO r
 handling_ :: Getter SomeException a     -> IO r -> IO r -> IO r
 handling_ :: Fold SomeException a       -> IO r -> IO r -> IO r

Throwing

throwing :: AReview s SomeException a b -> b -> a

Throw an Exception described by a Prism. Exceptions may be thrown from purely functional code, but may only be caught within the IO monad.

throwing l ≡ reviews l throw

 throwing :: Prism' SomeException t -> t -> a
 throwing :: Iso' SomeException t   -> t -> a

throwingIO :: AReview s SomeException a b -> b -> IO a

A variant of throwing that can only be used within the IO monad.

throwingIO l ≡ reviews l throwIO

Although throwingIO has a type that is a specialization of the type of throwing, the two functions are subtly different:

 throwing l e seq x   ≡ throwing e
 throwingIO l e seq x ≡ x

The first example will cause the exception e to be raised, whereas the second one won't. In fact, throwingIO will only cause an exception to be raised when it is used within the IO monad. The throwingIO variant should be used in preference to throwing to raise an exception within the IO monad because it guarantees ordering with respect to other IO operations, whereas throwing does not.

 throwingIO :: Prism' SomeException t -> t -> IO a
 throwingIO :: Iso' SomeException t   -> t -> IO a

throwingTo :: ThreadId -> AReview s SomeException a b -> b -> IO ()

throwingTo raises an exception specified by a Prism in the target thread

throwingTo thread l ≡ reviews l (throwTo thread)

 throwingTo :: ThreadId -> Prism' SomeException t -> t -> IO a
 throwingTo :: ThreadId -> Iso' SomeException t   -> t -> IO a

Exceptions

exception :: Exception a => Prism' SomeException a

Traverse the strongly typed Exception contained in SomeException where the type of your function matches the desired Exception.

 exception :: (Applicative f, Exception a, Exception b)
           => (a -> f b) -> SomeException -> f SomeException

IOExceptions

class AsIOException p f t where

Exceptions that occur in the IO monad. An IOException records a more specific error type, a descriptive string and maybe the handle that was used when the error was flagged.

Due to their richer structure relative to other exceptions, these have a more carefully overloaded signature.

Methods

_IOException :: Overloaded' p f t IOException

Unfortunately the name ioException is taken by base for throwing IOExceptions.

 _IOException :: Equality' IOException   IOException
 _IOException :: Prism' SomeException IOException

Many combinators for working with an IOException are available in System.IO.Error.Lens.

Arithmetic Exceptions

class AsArithException t where

Arithmetic exceptions.

Methods

_ArithException :: Prism' t ArithException

_Overflow :: AsArithException t => Prism' t ()

Handle arithmetic _Overflow.

_Overflow ≡ _ArithException . _Overflow

 _Overflow :: Prism' ArithException ArithException
 _Overflow :: Prism' SomeException  ArithException

_Underflow :: AsArithException t => Prism' t ()

Handle arithmetic _Underflow.

_Underflow ≡ _ArithException . _Underflow

 _Underflow :: Prism' ArithException ArithException
 _Underflow :: Prism' SomeException ArithException

_LossOfPrecision :: AsArithException t => Prism' t ()

Handle arithmetic loss of precision.

_LossOfPrecision ≡ _ArithException . _LossOfPrecision

 _LossOfPrecision :: Prism' ArithException ArithException
 _LossOfPrecision :: Prism' SomeException ArithException

_DivideByZero :: AsArithException t => Prism' t ()

Handle division by zero.

_DivideByZero ≡ _ArithException . _DivideByZero

_Denormal :: AsArithException t => Prism' t ()

Handle exceptional _Denormalized floating point.

_Denormal ≡ _ArithException . _Denormal

Array Exceptions

class AsArrayException t where

Exceptions generated by array operations

Methods

_ArrayException :: Prism' t ArrayException

Extract information about an array exception.

_IndexOutOfBounds :: AsArrayException t => Prism' t String

An attempt was made to index an array outside its declared bounds.

_IndexOutOfBounds ≡ _ArrayException . _IndexOutOfBounds

_UndefinedElement :: AsArrayException t => Prism' t String

An attempt was made to evaluate an element of an array that had not been initialized.

_UndefinedElement ≡ _ArrayException . _UndefinedElement

Assertion Failed

class AsAssertionFailed t where

assert was applied to False.

Methods

_AssertionFailed :: Prism' t String

>>> handling _AssertionFailed (\ xs -> "caught" <$ guard ("<interactive>" `isInfixOf` xs) ) $ assert False (return "uncaught")
"caught"

Async Exceptions

class AsAsyncException t where

Asynchronous exceptions.

Methods

_AsyncException :: Prism' t AsyncException

_StackOverflow :: AsAsyncException t => Prism' t ()

The current thread's stack exceeded its limit. Since an exception has been raised, the thread's stack will certainly be below its limit again, but the programmer should take remedial action immediately.

_HeapOverflow :: AsAsyncException t => Prism' t ()

The program's heap is reaching its limit, and the program should take action to reduce the amount of live data it has.

Notes:

It is undefined which thread receives this exception.

GHC currently does not throw HeapOverflow exceptions.

_ThreadKilled :: AsAsyncException t => Prism' t ()

This exception is raised by another thread calling killThread, or by the system if it needs to terminate the thread for some reason.

_UserInterrupt :: AsAsyncException t => Prism' t ()

This exception is raised by default in the main thread of the program when the user requests to terminate the program via the usual mechanism(s) (e.g. Control-C in the console).

Non-Termination

class AsNonTermination t where

Methods

_NonTermination :: Prism' t ()

Thrown when the runtime system detects that the computation is guaranteed not to terminate. Note that there is no guarantee that the runtime system will notice whether any given computation is guaranteed to terminate or not.

Nested Atomically

class AsNestedAtomically t where

Methods

_NestedAtomically :: Prism' t ()

Thrown when the program attempts to call atomically, from the stm package, inside another call to atomically.

Blocked Indefinitely

on MVar

class AsBlockedIndefinitelyOnMVar t where

Methods

_BlockedIndefinitelyOnMVar :: Prism' t ()

The thread is blocked on an MVar, but there are no other references to the MVar so it can't ever continue.

on STM

class AsBlockedIndefinitelyOnSTM t where

Methods

_BlockedIndefinitelyOnSTM :: Prism' t ()

The thread is waiting to retry an STM transaction, but there are no other references to any TVars involved, so it can't ever continue.

Deadlock

class AsDeadlock t where

Methods

_Deadlock :: Prism' t ()

There are no runnable threads, so the program is deadlocked. The Deadlock exception is raised in the main thread only.

No Such Method

class AsNoMethodError t where

Methods

_NoMethodError :: Prism' t String

A class method without a definition (neither a default definition, nor a definition in the appropriate instance) was called.

Pattern Match Failure

class AsPatternMatchFail t where

Methods

_PatternMatchFail :: Prism' t String

A pattern match failed.

Information about the source location of the pattern

Record

class AsRecConError t where

Methods

_RecConError :: Prism' t String

An uninitialised record field was used.

Information about the source location where the record was constructed

class AsRecSelError t where

A record selector was applied to a constructor without the appropriate field. This can only happen with a datatype with multiple constructors, where some fields are in one constructor but not another.

Methods

_RecSelError :: Prism' t String

Information about the source location where the record selection occurred

class AsRecUpdError t where

A record update was performed on a constructor without the appropriate field. This can only happen with a datatype with multiple constructors, where some fields are in one constructor but not another.

Methods

_RecUpdError :: Prism' t String

Information about the source location where the record was updated

Error Call

class AsErrorCall t where

This is thrown when the user calls error.

Methods

_ErrorCall :: Prism' t String

Retrieve the argument given to error.

ErrorCall is isomorphic to a String

>>> catching _ErrorCall (error "touch down!") return
"touch down!"