Control.Lens.Action

Description

Composable Actions

type Action m a c = forall f r. Effective m r f => (c -> f c) -> a -> f a

An Action is a Getter enriched with access to a Monad for side-effects.

Every Getter can be used as an Action

You can compose an Action with another Action using (.) from the Prelude.

act :: Monad m => (a -> m c) -> Action m a c

Construct an Action from a monadic side-effect

acts :: Action m (m a) a

A self-running Action, analogous to join.

acts = act id

>>> (1,"hello")^!_2.acts.to succ
"ifmmp"

perform :: Monad m => Acting m c a b c d -> a -> m c

Perform an Action.

 perform = flip (^!)

performs :: Monad m => Acting m e a b c d -> (c -> e) -> a -> m e

Perform an Action and modify the result.

liftAct :: (MonadTrans t, Monad m) => Acting m c a b c d -> Action (t m) a c

Apply a Monad transformer to an Action.

(^!) :: Monad m => a -> Acting m c a b c d -> m c

Perform an Action

>>> ["hello","world"]^!folded.act putStrLn
hello
world

Folds with Effecs

type MonadicFold m a c = forall f r. (Effective m r f, Applicative f) => (c -> f c) -> a -> f a

A MonadicFold is a Fold enriched with access to a Monad for side-effects.

Every Fold can be used as a MonadicFold, that simply ignores the access to the Monad.

You can compose a MonadicFold with another MonadicFold using (.) from the Prelude.

Implementation Details

type Acting m r a b c d = (c -> Effect m r d) -> a -> Effect m r b

Used to evaluate an Action.