Control.Lens.Action

Description

Composable Actions

type Action m s a = forall f r. Effective m r f => (a -> f a) -> s -> f s

An Action is a Getter enriched with access to a Monad for side-effects.

Every Getter can be used as an Action

You can compose an Action with another Action using (.) from the Prelude.

act :: Monad m => (s -> m a) -> Action m s a

Construct an Action from a monadic side-effect

acts :: Action m (m a) a

A self-running Action, analogous to join.

acts = act id

>>> (1,"hello")^!_2.acts.to succ
"ifmmp"

perform :: Monad m => Acting m a s t a b -> s -> m a

Perform an Action.

 perform = flip (^!)

performs :: Monad m => Acting m e s t a b -> (a -> e) -> s -> m e

Perform an Action and modify the result.

liftAct :: (MonadTrans trans, Monad m) => Acting m a s t a b -> Action (trans m) s a

Apply a Monad transformer to an Action.

(^!) :: Monad m => s -> Acting m a s t a b -> m a

Perform an Action

>>> ["hello","world"]^!folded.act putStrLn
hello
world

Folds with Effecs

type MonadicFold m s a = forall f r. (Effective m r f, Applicative f) => (a -> f a) -> s -> f s

A MonadicFold is a Fold enriched with access to a Monad for side-effects.

Every Fold can be used as a MonadicFold, that simply ignores the access to the Monad.

You can compose a MonadicFold with another MonadicFold using (.) from the Prelude.

Implementation Details

type Acting m r s t a b = (a -> Effect m r b) -> s -> Effect m r t

Used to evaluate an Action.