Control.Lens.At

Description

At

class At m where

At provides a Lens that can be used to read, write or delete the value associated with a key in a Map-like container on an ad hoc basis.

An instance of At should satisfy:

 ix k ≡ at k <. traverse

Methods

at :: Index m -> IndexedLens' (Index m) m (Maybe (IxValue m))

>>> Map.fromList [(1,"world")] ^.at 1
Just "world"

>>> at 1 ?~ "hello" $ Map.empty
fromList [(1,"hello")]

Note: Map-like containers form a reasonable instance, but not Array-like ones, where you cannot satisfy the Lens laws.

Ixed

type family IxValue m :: *

This provides a common notion of a value at an index that is shared by both Ixed and At.

class (Functor f, Contains (Accessor (IxValue m)) m) => Ixed f m where

This simple IndexedTraversal lets you traverse the value at a given key in a Map or element at an ordinal position in a list or Seq.

Methods

ix :: Index m -> IndexedLensLike' (Index m) f m (IxValue m)

This simple IndexedTraversal lets you traverse the value at a given key in a Map or element at an ordinal position in a list or Seq.

NB: Setting the value of this Traversal will only set the value in the Lens if it is already present.

If you want to be able to insert missing values, you want at.

>>> Seq.fromList [a,b,c,d] & ix 2 %~ f
fromList [a,b,f c,d]

>>> Seq.fromList [a,b,c,d] & ix 2 .~ e
fromList [a,b,e,d]

>>> Seq.fromList [a,b,c,d] ^? ix 2
Just c

>>> Seq.fromList [] ^? ix 2
Nothing

ixAt :: (Applicative f, At m) => Index m -> IndexedLensLike' (Index m) f m (IxValue m)

A definition of ix for types with an At instance. This is the default if you don't specify a definition for ix.

ixEach :: (Applicative f, Eq (Index m), Each f m m (IxValue m) (IxValue m)) => Index m -> IndexedLensLike' (Index m) f m (IxValue m)

A definition of ix for types with an Each instance.

Contains

class Functor f => Contains f m where

This class provides a simple IndexedFold (or IndexedTraversal) that lets you view (and modify) information about whether or not a container contains a given Index.

Methods

contains :: Index m -> IndexedLensLike' (Index m) f m Bool

>>> IntSet.fromList [1,2,3,4] ^. contains 3
True

>>> IntSet.fromList [1,2,3,4] ^. contains 5
False

>>> IntSet.fromList [1,2,3,4] & contains 3 .~ False
fromList [1,2,4]

containsIx :: (Gettable f, Ixed (Accessor Any) m) => Index m -> IndexedLensLike' (Index m) f m Bool

A definition of contains for types with an Ix instance.

containsAt :: (Gettable f, At m) => Index m -> IndexedLensLike' (Index m) f m Bool

A definition of ix for types with an At instance. This is the default if you don't specify a definition for contains and you are on GHC >= 7.0.2

containsLength :: forall i s. (Ord i, Num i) => (s -> i) -> i -> IndexedGetter i s Bool

Construct a contains check based on some notion of length for the container.

containsN :: Int -> Int -> IndexedGetter Int s Bool

Construct a contains check for a fixed number of elements.

containsTest :: forall i s. (i -> s -> Bool) -> i -> IndexedGetter i s Bool

Construct a contains check that uses an arbitrary test.

containsLookup :: forall i s a. (i -> s -> Maybe a) -> i -> IndexedGetter i s Bool

Construct a contains check that uses an arbitrary lookup function.

Deprecated

_at, resultAt :: Ixed f m => Index m -> IndexedLensLike' (Index m) f m (IxValue m)

Deprecated aliases for ix.