Control.Lens.At

Description

Indexed Lens

class At m where

At provides a lens that can be used to read, write or delete the value associated with a key in a map-like container on an ad hoc basis.

An instance of At should satisfy:

el k ≡ at k <. traverse

Methods

at :: IxKey m -> SimpleIndexedLens (IxKey m) m (Maybe (IxValue m))

>>> Map.fromList [(1,"world")] ^.at 1
Just "world"

>>> at 1 ?~ "hello" $ Map.empty
fromList [(1,"hello")]

Note: Map-like containers form a reasonable instance, but not Array-like ones, where you cannot satisfy the Lens laws.

Indexed Traversal

class Ixed f m where

This simple indexed traversal lets you traverse the value at a given key in a map or element at an ordinal position in a list or sequence.

Methods

ix :: Indexable (IxKey m) k => IxKey m -> SimpleOverloaded k f m (IxValue m)

What is the index type? | This simple indexed traversal lets you traverse the value at a given key in a map.

NB:* _setting_ the value of this Traversal will only set the value in the lens if it is already present.

If you want to be able to insert missing values, you want at.

>>> Seq.fromList [a,b,c,d] & ix 2 %~ f
fromList [a,b,f c,d]

>>> Seq.fromList [a,b,c,d] & ix 2 .~ e
fromList [a,b,e,d]

>>> Seq.fromList [a,b,c,d] ^? ix 2
Just c

>>> Seq.fromList [] ^? ix 2
Nothing

>>> IntSet.fromList [1,2,3,4] & ix 3 .~ False
fromList [1,2,4]

>>> IntSet.fromList [1,2,3,4] ^. ix 3
True

Deprecated

_at :: (Indexable (IxKey m) k, Ixed f m) => IxKey m -> SimpleOverloaded k f m (IxValue m)

Deprecated: use 'ix'. This function will be removed in version 3.9

A deprecated alias for ix

contains :: (Indexable (IxKey m) k, Ixed f m) => IxKey m -> SimpleOverloaded k f m (IxValue m)

Deprecated: use 'ix'. This function will be removed in version 3.9

A deprecated alias for ix

resultAt :: (Indexable (IxKey m) k, Ixed f m) => IxKey m -> SimpleOverloaded k f m (IxValue m)

Deprecated: use 'ix'. This function will be removed in version 3.9

A deprecated alias for ix