Control.Lens.Classes

Description

Getters

class Functor f => Gettable f where

Generalizing Const so we can apply simple Applicative transformations to it and so we can get nicer error messages

A Gettable Functor ignores its argument, which it carries solely as a phantom type parameter.

To ensure this, an instance of Gettable is required to satisfy:

id = fmap f = coerce

Which is equivalent to making a Gettable f an "anyvariant" functor.

Methods

coerce :: f a -> f b

Replace the phantom type argument.

noEffect :: (Applicative f, Gettable f) => f a

The mempty equivalent for a Gettable Applicative Functor.

Actions

class (Monad m, Gettable f) => Effective m r f | f -> m r where

An Effective Functor ignores its argument and is isomorphic to a Monad wrapped around a value.

That said, the Monad is possibly rather unrelated to any Applicative structure.

Methods

effective :: m r -> f a

ineffective :: f a -> m r

Setters

class Applicative f => Settable f where

Anything Settable must be isomorphic to the Identity Functor.

Methods

untainted :: f a -> a

untaintedDot :: (a -> f b) -> a -> b

taintedDot :: (a -> b) -> a -> f b

Costrong

class Functor f => Costrong f where

Methods

costrength :: f (Either a b) -> Either a (f b)

Algebraics

class (Profunctor k, Functor g) => Algebraic g k | k -> g where

Used to provide overloading of symmetric lenses as regular lenses.

There are two relevant instances: (->) (for using an 'Iso'/'Prism' as a regular lens) and Cokleisli g (for using it as a symmetric lens).

Methods

algebraically :: (k a b -> k s t) -> (g a -> b) -> g s -> t

unalgebraically :: ((g a -> b) -> g s -> t) -> k a b -> k s t

Indexable

class Profunctor c => Indexable i c where

This class permits overloading of function application for things that also admit a notion of a key or index.

Methods

indexed :: c a b -> i -> a -> b

Build a function from an Indexed function