Control.Lens.Internal

Description

These are some of the explicit Functor instances that leak into the type signatures of Control.Lens. You shouldn't need to import this module directly, unless you are coming up with a whole new kind of "Family" and need to add instances.

Internal Types

newtype May a

Make a monoid out of Maybe for error handling

Constructors

May

Fields

getMay :: Maybe a

newtype Folding f a

A Monoid for a Gettable Applicative.

Constructors

Folding

Fields

getFolding :: f a

newtype Effect m r a

Wrap a monadic effect with a phantom type argument.

Constructors

Effect

Fields

getEffect :: m r

newtype EffectRWS w st m s a

Wrap a monadic effect with a phantom type argument. Used when magnifying RWST.

Constructors

EffectRWS

Fields

getEffectRWS :: st -> m (s, st, w)

newtype Accessor r a

Used instead of Const to report

No instance of (Settable Accessor)

when the user attempts to misuse a Setter as a Getter, rather than a monolithic unification error.

Constructors

Accessor

Fields

runAccessor :: r

newtype Err e a

Make a monoid out of Either for error handling

Constructors

Err

Fields

getErr :: Either e a

newtype Traversed f

Used internally by traverseOf_ and the like.

Constructors

Traversed

Fields

getTraversed :: f ()

newtype Sequenced m

Used internally by mapM_ and the like.

Constructors

Sequenced

Fields

getSequenced :: m ()

newtype Focusing m s a

Used by Zoom to zoom into StateT

Constructors

Focusing

Fields

unfocusing :: m (s, a)

newtype FocusingWith w m s a

Used by Zoom to zoom into RWST

Constructors

FocusingWith

Fields

unfocusingWith :: m (s, a, w)

newtype FocusingPlus w k s a

Used by Zoom to zoom into WriterT.

Constructors

FocusingPlus

Fields

unfocusingPlus :: k (s, w) a

newtype FocusingOn f k s a

Used by Zoom to zoom into MaybeT or ListT

Constructors

FocusingOn

Fields

unfocusingOn :: k (f s) a

newtype FocusingMay k s a

Used by Zoom to zoom into ErrorT

Constructors

FocusingMay

Fields

unfocusingMay :: k (May s) a

newtype FocusingErr e k s a

Used by Zoom to zoom into ErrorT

Constructors

FocusingErr

Fields

unfocusingErr :: k (Err e s) a

newtype Mutator a

Mutator is just a renamed Identity functor to give better error messages when someone attempts to use a getter as a setter.

Most user code will never need to see this type.

Constructors

Mutator

Fields

runMutator :: a

newtype Bazaar a b t

This is used to characterize a Traversal.

a.k.a. indexed Cartesian store comonad, indexed Kleene store comonad, or an indexed FunList.

http://twanvl.nl/blog/haskell/non-regular1

Bazaar a b t is isomorphic to data Bazaar a b t = Buy t | Trade (Bazaar a b (b -> t)) a, and to exists s. (s, Traversal s t a b).

A Bazaar is like a Traversal that has already been applied to some structure.

Where a Context a b t holds an a and a function from b to t, a Bazaar a b t holds N as and a function from N bs to t.

Mnemonically, a Bazaar holds many stores and you can easily add more.

This is a final encoding of Bazaar.

Constructors

Bazaar

Fields

runBazaar :: forall f. Applicative f => (a -> f b) -> f t

bazaar :: Applicative f => (a -> f b) -> Bazaar a b t -> f t

Given an action to run for each matched pair, traverse a bazaar.

bazaar :: Traversal (Bazaar a b t) t a b

duplicateBazaar :: Bazaar a c t -> Bazaar a b (Bazaar b c t)

Bazaar is an indexed Comonad.

sell :: a -> Bazaar a b b

A trivial Bazaar.

newtype BazaarT a b g t

BazaarT is like Bazaar, except that it provides a questionable Gettable instance To protect this instance it relies on the soundness of another Gettable type, and usage conventions.

For example. This lets us write a suitably polymorphic and lazy taking, but there must be a better way!

Constructors

BazaarT

Fields

runBazaarT :: forall f. Applicative f => (a -> f b) -> f t

bazaarT :: Applicative f => (a -> f b) -> BazaarT a b g t -> f t

Extract from a BazaarT.

bazaarT = flip runBazaarT

duplicateBazaarT :: BazaarT a c f t -> BazaarT a b f (BazaarT b c f t)

BazaarT is an indexed Comonad.

sellT :: a -> BazaarT a b f b

A trivial BazaarT.

data Context a b t

The indexed store can be used to characterize a Lens and is used by clone

Context a b t is isomorphic to newtype Context a b t = Context { runContext :: forall f. Functor f => (a -> f b) -> f t }, and to exists s. (s, Lens s t a b).

A Context is like a Lens that has already been applied to a some structure.

Constructors

Context (b -> t) a

data Max a

Used for maximumOf

Constructors

NoMax

Max a

getMax :: Max a -> Maybe a

Obtain the maximum

data Min a

Used for minimumOf

Constructors

NoMin

Min a

getMin :: Min a -> Maybe a

Obtain the minimum.

newtype Indexing f a

Applicative composition of State Int with a Functor, used by indexed

Constructors

Indexing

Fields

runIndexing :: Int -> (f a, Int)

Overloadings

data Projected x y where

This data type is used to capture all of the information provided by the Projective class, so you can turn a Projection around into a Getter or otherwise muck around with its internals.

If you see a function that expects a Project, it is probably just expecting a Projection.

Constructors

Projected :: (b -> t) -> (s -> Either t a) -> Projected (a -> f b) (s -> f t)

ProjectedId :: Projected x x

data Isomorphism x y where

Reify all of the information given to you by being Isomorphic.

Constructors

Isomorphism :: (s -> a) -> (b -> t) -> Isomorphism (a -> f b) (s -> f t)

IsomorphismId :: Isomorphism x x

newtype Indexed i a b

A function with access to a index. This constructor may be useful when you need to store a Indexable in a container to avoid ImpredicativeTypes.

Constructors

Indexed

Fields

withIndex :: (i -> a) -> b