Control.Lens.Internal

Description

These are some of the explicit Functor instances that leak into the type signatures of Control.Lens. You shouldn't need to import this module directly, unless you are coming up with a whole new kind of "Family" and need to add instances.

This module also includes orphan instances for ((,) e) that should be supplied by base.

Internal Classes

Getters

class Functor f => Gettable f where

Generalizing Const so we can apply simple Applicative transformations to it and so we can get nicer error messages

A Gettable Functor ignores its argument, which it carries solely as a phantom type parameter.

To ensure this, an instance of Gettable is required to satisfy:

id = fmap f = coerce

Which is equivalent to making a Gettable f an "anyvariant" functor.

Methods

coerce :: f a -> f b

Replace the phantom type argument.

noEffect :: (Applicative f, Gettable f) => f a

The mempty equivalent for a Gettable Applicative Functor.

Actions

class (Monad m, Gettable f) => Effective m r f | f -> m r where

An Effective Functor ignores its argument and is isomorphic to a Monad wrapped around a value.

That said, the Monad is possibly rather unrelated to any Applicative structure.

Methods

effective :: m r -> f a

ineffective :: f a -> m r

Setters

class (Applicative f, Distributive f, Traversable f) => Settable f where

Anything Settable must be isomorphic to the Identity Functor.

Methods

untainted :: f a -> a

untaintedDot :: Profunctor p => p a (f b) -> p a b

taintedDot :: Profunctor p => p a b -> p a (f b)

Reviewable

class Profunctor p => Reviewable p where

Methods

retagged :: p a b -> p s b

Indexable

class (Profunctor p, Prismatic p, Lenticular p, Corepresentable p, Comonad (Corep p), Traversable (Corep p), Representable p, Monad (Rep p), Distributive (Rep p), ArrowLoop p, ArrowApply p, ArrowChoice p) => SelfAdjoint p where

Methods

distrib :: Functor f => p a b -> p (f a) (f b)

class SelfAdjoint p => Indexable i p where

This class permits overloading of function application for things that also admit a notion of a key or index.

Methods

indexed :: p a b -> i -> a -> b

Build a function from an Indexed function.

Indexed Functors

class IndexedFunctor w where

Methods

ifmap :: (s -> t) -> w a b s -> w a b t

Indexed Comonads

class IndexedFunctor w => IndexedComonad w where

Methods

iextract :: w a a t -> t

iduplicate :: w a c t -> w a b (w b c t)

iextend :: (w b c t -> r) -> w a c t -> w a b r

Indexed Store Comonad

class IndexedComonad w => IndexedComonadStore w where

Methods

ipos :: w a c t -> a

ipeek :: c -> w a c t -> t

ipeeks :: (a -> c) -> w a c t -> t

iseek :: b -> w a c t -> w b c t

iseeks :: (a -> b) -> w a c t -> w b c t

iexperiment :: Functor f => (b -> f c) -> w b c t -> f t

context :: w a b t -> Context a b t

Internal Types

newtype May a

Make a monoid out of Maybe for error handling

Constructors

May

Fields

getMay :: Maybe a

newtype Effect m r a

Wrap a monadic effect with a phantom type argument.

Constructors

Effect

Fields

getEffect :: m r

newtype EffectRWS w st m s a

Wrap a monadic effect with a phantom type argument. Used when magnifying RWST.

Constructors

EffectRWS

Fields

getEffectRWS :: st -> m (s, st, w)

newtype Err e a

Make a monoid out of Either for error handling

Constructors

Err

Fields

getErr :: Either e a

newtype Traversed f

Used internally by traverseOf_ and the like.

Constructors

Traversed

Fields

getTraversed :: f ()

newtype Sequenced m

Used internally by mapM_ and the like.

Constructors

Sequenced

Fields

getSequenced :: m ()

newtype Focusing m s a

Used by Zoom to zoom into StateT

Constructors

Focusing

Fields

unfocusing :: m (s, a)

newtype FocusingWith w m s a

Used by Zoom to zoom into RWST

Constructors

FocusingWith

Fields

unfocusingWith :: m (s, a, w)

newtype FocusingPlus w k s a

Used by Zoom to zoom into WriterT.

Constructors

FocusingPlus

Fields

unfocusingPlus :: k (s, w) a

newtype FocusingOn f k s a

Used by Zoom to zoom into MaybeT or ListT

Constructors

FocusingOn

Fields

unfocusingOn :: k (f s) a

newtype FocusingMay k s a

Used by Zoom to zoom into ErrorT

Constructors

FocusingMay

Fields

unfocusingMay :: k (May s) a

newtype FocusingErr e k s a

Used by Zoom to zoom into ErrorT

Constructors

FocusingErr

Fields

unfocusingErr :: k (Err e s) a

newtype Folding f a

A Monoid for a Gettable Applicative.

Constructors

Folding

Fields

getFolding :: f a

data Max a

Used for maximumOf

Constructors

NoMax

Max a

getMax :: Max a -> Maybe a

Obtain the maximum

data Min a

Used for minimumOf

Constructors

NoMin

Min a

getMin :: Min a -> Maybe a

Obtain the minimum.

newtype Indexing f a

Applicative composition of State Int with a Functor, used by indexed

Constructors

Indexing

Fields

runIndexing :: Int -> (Int, f a)

indexing :: Indexable Int p => ((a -> Indexing f b) -> s -> Indexing f t) -> p a (f b) -> s -> f t

Transform a Traversal into an IndexedTraversal or a Fold into an IndexedFold, etc.

 indexing :: Traversal s t a b -> IndexedTraversal Int s t a b
 indexing :: Prism s t a b     -> IndexedTraversal Int s t a b
 indexing :: Lens s t a b      -> IndexedLens Int  s t a b
 indexing :: Iso s t a b       -> IndexedLens Int s t a b
 indexing :: Fold s a          -> IndexedFold Int s a
 indexing :: Getter s a        -> IndexedGetter Int s a

indexing :: Indexable Int p => LensLike (Indexing f) s t a b -> Overloading p (->) f s t a b

newtype Indexing64 f a

Applicative composition of State Int64 with a Functor, used by indexed64

Constructors

Indexing64

Fields

runIndexing64 :: Int64 -> (Int64, f a)

indexing64 :: Indexable Int64 p => ((a -> Indexing64 f b) -> s -> Indexing64 f t) -> p a (f b) -> s -> f t

Transform a Traversal into an IndexedTraversal or a Fold into an IndexedFold, etc.

This combinator is like indexing except that it handles large Traversals and Folds gracefully.

 indexing64 :: Traversal s t a b -> IndexedTraversal Int64 s t a b
 indexing64 :: Prism s t a b     -> IndexedTraversal Int64 s t a b
 indexing64 :: Lens s t a b      -> IndexedLens Int64 s t a b
 indexing64 :: Iso s t a b       -> IndexedLens Int64 s t a b
 indexing64 :: Fold s a          -> IndexedFold Int64 s a
 indexing64 :: Getter s a        -> IndexedGetter Int64 s a

indexing64 :: Indexable Int64 p => LensLike (Indexing64 f) s t a b -> Overloading p (->) f s t a b

Common Types

newtype Accessor r a

Used instead of Const to report

No instance of (Settable Accessor)

when the user attempts to misuse a Setter as a Getter, rather than a monolithic unification error.

Constructors

Accessor

Fields

runAccessor :: r

newtype Mutator a

Mutator is just a renamed Identity functor to give better error messages when someone attempts to use a getter as a setter.

Most user code will never need to see this type.

Constructors

Mutator

Fields

runMutator :: a

newtype Reviewed a b

Constructors

Reviewed

Fields

runReviewed :: b

newtype Exchange a b s t

Constructors

Exchange

Fields

runExchange :: (s -> a, b -> t)

newtype Market a b s t

Constructors

Market

Fields

runMarket :: (b -> t, s -> Either t a)

type Market' a = Market a a

type Market' a s t = Market a a s t

data Identical a b s t where

Constructors

Identical :: Identical a b a b

newtype Indexed i a b

A function with access to a index. This constructor may be useful when you need to store an Indexable in a container to avoid ImpredicativeTypes.

index :: Indexed i a b -> i -> a -> b

Constructors

Indexed

Fields

runIndexed :: i -> a -> b

data Context a b t

The indexed store can be used to characterize a Lens and is used by clone.

Context a b t is isomorphic to newtype Context a b t = Context { runContext :: forall f. Functor f => (a -> f b) -> f t }, and to exists s. (s, Lens s t a b).

A Context is like a Lens that has already been applied to a some structure.

Constructors

Context (b -> t) a

type Context' a = Context a a

type Context' a s = Context a a s

newtype Pretext p q a b t

Constructors

Pretext

Fields

runPretext :: forall f. Functor f => p a (f b) `q` f t

type Pretext' p q a = Pretext p q a a

type Pretext' p q a s = Pretext p q a a s

newtype PretextT p q g a b t

Constructors

PretextT

Fields

runPretextT :: forall f. Functor f => p a (f b) `q` f t

type PretextT' p q g a = PretextT p q g a a

type PretextT' p q g a s = PretextT p q g a a s

class (Profunctor p, Profunctor q) => Bizarre p q w | w -> p q where

Methods

bazaar :: Applicative f => p a (f b) -> q (w a b t) (f t)

newtype Bazaar p q a b t

This is used to characterize a Traversal.

a.k.a. indexed Cartesian store comonad, indexed Kleene store comonad, or an indexed FunList.

http://twanvl.nl/blog/haskell/non-regular1

Bazaar a b t is isomorphic to data Bazaar a b t = Buy t | Trade (Bazaar a b (b -> t)) a, and to exists s. (s, Traversal s t a b).

A Bazaar is like a Traversal that has already been applied to some structure.

Where a Context a b t holds an a and a function from b to t, a Bazaar a b t holds N as and a function from N bs to t.

Mnemonically, a Bazaar holds many stores and you can easily add more.

This is a final encoding of Bazaar. This is a final encoding of Bazaar.

Constructors

Bazaar

Fields

runBazaar :: forall f. Applicative f => p a (f b) `q` f t

type Bazaar' p q a = Bazaar p q a a

newtype BazaarT p q g a b t

BazaarT is like Bazaar, except that it provides a questionable Gettable instance To protect this instance it relies on the soundness of another Gettable type, and usage conventions.

For example. This lets us write a suitably polymorphic and lazy taking, but there must be a better way!

Constructors

BazaarT

Fields

runBazaarT :: forall f. Applicative f => p a (f b) `q` f t

type BazaarT' p q g a = BazaarT p q g a a

class Corepresentable p => Sellable p k | k -> p where

Methods

sell :: p a (k a b b)

data Level i a

Constructors

Two !Word !(Level i a) !(Level i a)

One i a

Zero

newtype Deepening i a

Constructors

Deepening

Fields

runDeepening :: forall r. Int -> (Level i a -> Bool -> r) -> r

deepening :: i -> a -> Deepening i a

newtype Flows i b a

Constructors

Flows

Fields

runFlows :: [Level i b] -> a