Control.Lens.Internal

Description

These are some of the explicit Functor instances that leak into the type signatures of Control.Lens. You shouldn't need to import this module directly, unless you are coming up with a whole new kind of "Family" and need to add instances.

Implementation details

data IndexedStore c d a

The indexed store can be used to characterize a Lens and is used by clone

Constructors

IndexedStore (d -> a) c

newtype Focusing m c a

Used by Focus

Constructors

Focusing

Fields

unfocusing :: m (c, a)

newtype Traversed f

Used internally by traverseOf_ and the like.

Constructors

Traversed

Fields

getTraversed :: f ()

newtype Sequenced m

Used internally by mapM_ and the like.

Constructors

Sequenced

Fields

getSequenced :: m ()

newtype AppliedState f a

Applicative composition of State Int with a Functor, used by elementOf, elementsOf, traverseElement, traverseElementsOf

Constructors

AppliedState

Fields

runAppliedState :: Int -> (f a, Int)

data Min a

Used for minimumOf

Constructors

NoMin

Min a

getMin :: Min a -> Maybe a

Obtain the minimum.

data Max a

Used for maximumOf

Constructors

NoMax

Max a

getMax :: Max a -> Maybe a

Obtain the maximum

newtype ElementOf f a

Used to find the nth element of a Traversal.

Constructors

ElementOf

Fields

getElementOf :: Int -> ElementOfResult f a

data ElementOfResult f a

The result of trying to find the nth element of a Traversal.

Constructors

Searching !Int a

Found !Int (f a)

NotFound String