Control.Lens.Internal

Description

These are some of the explicit Functor instances that leak into the type signatures of Control.Lens. You shouldn't need to import this module directly, unless you are coming up with a whole new kind of "Family" and need to add instances.

Implementation details

data Context c d a

The indexed store can be used to characterize a Lens and is used by clone

Constructors

Context (d -> a) c

newtype Focusing m c a

Used by Zoom to zoom into StateT

Constructors

Focusing

Fields

unfocusing :: m (c, a)

newtype FocusingWith w m c a

Used by Zoom to zoom into RWST

Constructors

FocusingWith

Fields

unfocusingWith :: m (c, a, w)

newtype FocusingPlus w k c a

Used by Zoom to zoom into WriterT.

Constructors

FocusingPlus

Fields

unfocusingPlus :: k (c, w) a

newtype FocusingOn f k c a

Used by Zoom to zoom into MaybeT or ListT

Constructors

FocusingOn

Fields

unfocusingOn :: k (f c) a

newtype FocusingErr e k c a

Used by Zoom to zoom into ErrorT

Constructors

FocusingErr

Fields

unfocusingErr :: k (Err e c) a

newtype Err e a

Make a monoid out of Either for error handling

Constructors

Err

Fields

getErr :: Either e a

newtype FocusingMay k c a

Used by Zoom to zoom into ErrorT

Constructors

FocusingMay

Fields

unfocusingMay :: k (May c) a

newtype May a

Make a monoid out of Maybe for error handling

Constructors

May

Fields

getMay :: Maybe a

newtype Traversed f

Used internally by traverseOf_ and the like.

Constructors

Traversed

Fields

getTraversed :: f ()

newtype Sequenced m

Used internally by mapM_ and the like.

Constructors

Sequenced

Fields

getSequenced :: m ()

newtype Indexing f a

Applicative composition of State Int with a Functor, used by indexed

Constructors

Indexing

Fields

runIndexing :: Int -> IndexingResult f a

data IndexingResult f a

The result of Indexing

Constructors

IndexingResult (f a) !Int

data Min a

Used for minimumOf

Constructors

NoMin

Min a

getMin :: Min a -> Maybe a

Obtain the minimum.

data Max a

Used for maximumOf

Constructors

NoMax

Max a

getMax :: Max a -> Maybe a

Obtain the maximum

newtype ElementOf f a

Used to find the nth element of a Traversal.

Constructors

ElementOf

Fields

getElementOf :: Int -> ElementOfResult f a

data ElementOfResult f a

The result of trying to find the nth element of a Traversal.

Constructors

Searching !Int a

Found !Int (f a)

NotFound String

newtype Bazaar c d a

This is used to characterize a Traversal.

a.k.a. indexed Cartesian store comonad, indexed Kleene store comonad, or an indexed FunList.

http://twanvl.nl/blog/haskell/non-regular1

Mnemonically, a Bazaar holds many stores and you can easily add more.

This is a final encoding of Bazaar.

Constructors

Bazaar

Fields

_runBazaar :: forall f. Applicative f => (c -> f d) -> f a

bazaar :: Applicative f => (c -> f d) -> Bazaar c d b -> f b

Given an action to run for each matched pair, traverse a bazaar.

duplicateBazaar :: Bazaar c e a -> Bazaar c d (Bazaar d e a)

Bazaar is an indexed Comonad.

sell :: c -> Bazaar c d d

A trivial Bazaar.

newtype Effect m r a

Wrap a monadic effect with a phantom type argument.

Constructors

Effect

Fields

getEffect :: m r

newtype EffectRWS w s m c a

Wrap a monadic effect with a phantom type argument. Used when magnifying RWST.

Constructors

EffectRWS

Fields

getEffectRWS :: s -> m (c, s, w)

class Functor f => Gettable f where

Generalizing Const so we can apply simple Applicative transformations to it and so we can get nicer error messages

A Gettable Functor ignores its argument, which it carries solely as a phantom type parameter.

To ensure this, an instance of Gettable is required to satisfy:

id = fmap f = coerce

Methods

coerce :: f a -> f b

Replace the phantom type argument.

newtype Accessor r a

Used instead of Const to report

No instance of (Settable Accessor)

when the user attempts to misuse a Setter as a Getter, rather than a monolithic unification error.

Constructors

Accessor

Fields

runAccessor :: r

class (Monad m, Gettable f) => Effective m r f | f -> m r where

An Effective Functor ignores its argument and is isomorphic to a monad wrapped around a value.

That said, the monad is possibly rather unrelated to any Applicative structure.

Methods

effective :: Isomorphic k => k (m r) (f a)

ineffective :: Effective m r f => Isomorphic k => k (f a) (m r)

A convenient antonym that is used internally.

class Applicative f => Settable f where

Anything Settable must be isomorphic to the Identity Functor.

Methods

untainted :: f a -> a

newtype Mutator a

Mutator is just a renamed Identity functor to give better error messages when someone attempts to use a getter as a setter.

Most user code will never need to see this type.

Constructors

Mutator

Fields

runMutator :: a