Control.Lens.Internal

Description

These are some of the explicit Functor instances that leak into the type signatures of Control.Lens. You shouldn't need to import this module directly for most use-cases.

Documentation

module Control.Lens.Internal.Action

module Control.Lens.Internal.Bazaar

module Control.Lens.Internal.Context

module Control.Lens.Internal.Fold

module Control.Lens.Internal.Getter

module Control.Lens.Internal.Indexed

module Control.Lens.Internal.Iso

module Control.Lens.Internal.Level

module Control.Lens.Internal.Magma

module Control.Lens.Internal.Prism

module Control.Lens.Internal.Review

module Control.Lens.Internal.Setter

module Control.Lens.Internal.Zoom