Control.Lens.Internal.Coerce

Description

This module provides a shim around coerce that defaults to unsafeCoerce on GHC < 7.8. It also exposes a type-restricted version of '(#.)' that works around a bizarre GHC 7.10–specific bug.

Documentation

coerce :: Coercible a b => a -> b

The function coerce allows you to safely convert between values of types that have the same representation with no run-time overhead. In the simplest case you can use it instead of a newtype constructor, to go from the newtype's concrete type to the abstract type. But it also works in more complicated settings, e.g. converting a list of newtypes to a list of concrete types.

coerce' :: forall a b. Coercible a b => b -> a

(#..) :: (Profunctor p, Coercible c b) => (b -> c) -> p a b -> p a c