Control.Lens.Internal.Getter

Description

Internal Classes

class (Contravariant f, Functor f) => Gettable f

This class is provided mostly for backwards compatibility with lens 3.8, but it can also shorten type signatures.

Getters

coerce :: (Contravariant f, Functor f) => f a -> f b

This Generalizes Const so we can apply simple Applicative transformations to it and so we can get nicer error messages.

A Functor you can coerce ignores its argument, which it carries solely as a phantom type parameter.

By the Functor and Contravariant laws, an instance of Gettable will necessarily satisfy:

id = fmap f = coerce = contramap g

noEffect :: (Contravariant f, Applicative f) => f a

The mempty equivalent for a Gettable Applicative Functor.

newtype AlongsideLeft f b a

Constructors

AlongsideLeft

Fields

getAlongsideLeft :: f (a, b)

newtype AlongsideRight f a b

Constructors

AlongsideRight

Fields

getAlongsideRight :: f (a, b)