Control.Lens.Internal.Indexed

Description

An Indexed Profunctor

newtype Indexed i a b

A function with access to a index. This constructor may be useful when you need to store an Indexable in a container to avoid ImpredicativeTypes.

index :: Indexed i a b -> i -> a -> b

Constructors

Indexed

Fields

runIndexed :: i -> a -> b

Classes

class (Profunctor p, Choice p, Strong p, Corepresentable p, Comonad (Corep p), Traversable (Corep p), Representable p, Monad (Rep p), MonadFix (Rep p), Distributive (Rep p), ArrowLoop p, ArrowApply p, ArrowChoice p) => Conjoined p where

Methods

distrib :: Functor f => p a b -> p (f a) (f b)

class Conjoined p => Indexable i p where

This class permits overloading of function application for things that also admit a notion of a key or index.

Methods

indexed :: p a b -> i -> a -> b

Build a function from an indexed function.

Indexing

newtype Indexing f a

Applicative composition of State Int with a Functor, used by indexed.

Constructors

Indexing

Fields

runIndexing :: Int -> (Int, f a)

indexing :: Indexable Int p => ((a -> Indexing f b) -> s -> Indexing f t) -> p a (f b) -> s -> f t

Transform a Traversal into an IndexedTraversal or a Fold into an IndexedFold, etc.

 indexing :: Traversal s t a b -> IndexedTraversal Int s t a b
 indexing :: Prism s t a b     -> IndexedTraversal Int s t a b
 indexing :: Lens s t a b      -> IndexedLens Int  s t a b
 indexing :: Iso s t a b       -> IndexedLens Int s t a b
 indexing :: Fold s a          -> IndexedFold Int s a
 indexing :: Getter s a        -> IndexedGetter Int s a

indexing :: Indexable Int p => LensLike (Indexing f) s t a b -> Overloading p (->) f s t a b

64-bit Indexing

newtype Indexing64 f a

Applicative composition of State Int64 with a Functor, used by indexed64.

Constructors

Indexing64

Fields

runIndexing64 :: Int64 -> (Int64, f a)

indexing64 :: Indexable Int64 p => ((a -> Indexing64 f b) -> s -> Indexing64 f t) -> p a (f b) -> s -> f t

Transform a Traversal into an IndexedTraversal or a Fold into an IndexedFold, etc.

This combinator is like indexing except that it handles large traversals and folds gracefully.

 indexing64 :: Traversal s t a b -> IndexedTraversal Int64 s t a b
 indexing64 :: Prism s t a b     -> IndexedTraversal Int64 s t a b
 indexing64 :: Lens s t a b      -> IndexedLens Int64 s t a b
 indexing64 :: Iso s t a b       -> IndexedLens Int64 s t a b
 indexing64 :: Fold s a          -> IndexedFold Int64 s a
 indexing64 :: Getter s a        -> IndexedGetter Int64 s a

indexing64 :: Indexable Int64 p => LensLike (Indexing64 f) s t a b -> Overloading p (->) f s t a b