Control.Lens.Internal.Iso

Description

Documentation

data Exchange a b s t

This is used internally by the Iso code to provide efficient access to the two functions that make up an isomorphism.

Constructors

Exchange (s -> a) (b -> t)