Control.Lens.Internal.Level

Description

This module provides implementation details of the combinators in Control.Lens.Level, which provides for the breadth-first Traversal of an arbitrary Traversal.

Levels

data Level i a

This data type represents a path-compressed copy of one level of a source data structure. We can safely use path-compression because we know the depth of the tree.

Path compression is performed by viewing a Level as a PATRICIA trie of the paths into the structure to leaves at a given depth, similar in many ways to a IntMap, but unlike a regular PATRICIA trie we do not need to store the mask bits merely the depth of the fork.

One invariant of this structure is that underneath a Two node you will not find any Zero nodes, so Zero can only occur at the root.

Constructors

Two !Word !(Level i a) !(Level i a)

One i a

Zero

newtype Deepening i a

This is an illegal Monoid used to construct a single Level.

Constructors

Deepening

Fields

runDeepening :: forall r. Int -> (Level i a -> Bool -> r) -> r

deepening :: i -> a -> Deepening i a

Generate the leaf of a given Deepening based on whether or not we're at the correct depth.

newtype Flows i b a

This is an illegal Applicative used to replace the contents of a list of consecutive Level values representing each layer of a structure into the original shape that they were derived from.

Attempting to Flow something back into a shape other than the one it was taken from will fail.

Constructors

Flows

Fields

runFlows :: [Level i b] -> a