Control.Lens.Internal.Prism

Description

Documentation

data Market a b s t

This type is used internally by the Prism code to provide efficient access to the two parts of a Prism.

Constructors

Market (b -> t) (s -> Either t a)

type Market' a = Market a a

type Market' a s t = Market a a s t