Control.Lens.Internal.Review

Description

Internal Classes

class (Profunctor p, Bifunctor p) => Reviewable p

This class is provided mostly for backwards compatibility with lens 3.8, but it can also shorten type signatures.

Reviews

retagged :: (Profunctor p, Bifunctor p) => p a b -> p s b

This is a profunctor used internally to implement Review

It plays a role similar to that of Accessor or Const do for Control.Lens.Getter

newtype Reviewed a b

Constructors

Reviewed

Fields

runReviewed :: b