Control.Lens.Internal.Zipper

Description

This module provides internal types and functions used in the implementation of Control.Lens.Zipper. You shouldn't need to import it directly, and the exported types can be used to break zipper invariants.

Documentation

>>> import Control.Lens

Zippers

data Top

This is used to represent the Top of the zipper.

Every zipper starts with Top.

e.g. Top :> a is the trivial zipper.

data h :> a

This is the type of a zipper. It visually resembles a "breadcrumb trail" as used in website navigation. Each breadcrumb in the trail represents a level you can move up to.

This type operator associates to the left, so you can use a type like

Top :> (String,Double) :> String :> Char

to represent a zipper from (String,Double) down to Char that has an intermediate crumb for the String containing the Char.

You can construct a zipper into *any* data structure with zipper.

>>> :t zipper (Just "hello")
zipper (Just "hello") :: Top :> Maybe [Char]

You can repackage up the contents of a zipper with rezip.

>>> rezip $ zipper 42
42

The combinators in this module provide lot of things you can do to the zipper while you have it open.

Note that a value of type h :> s :> a doesn't actually contain a value of type h :> s -- as we descend into a level, the previous level is unpacked and stored in Coil form. Only one value of type _ :> _ exists at any particular time for any particular zipper.

Constructors

Zipper (Coil h a) !Int [a] a [a]

type family Zipped h a

This represents the type a zipper will have when it is fully Zipped back up.

data Coil where

A Coil is a linked list of the levels above the current one. The length of a Coil is known at compile time.

This is part of the internal structure of a zipper. You shouldn't need to manipulate this directly.

Constructors

Coil :: Coil Top a

Snoc :: Coil h s -> SimpleLensLike (Bazaar a a) s a -> !Int -> [s] -> ([a] -> s) -> [s] -> Coil (h :> s) a

focus :: SimpleIndexedLens (Tape (h :> a)) (h :> a) a

This Lens views the current target of the zipper.

A Tape that can be used to get to the current location is available as the index of this Lens.

zipper :: a -> Top :> a

Construct a zipper that can explore anything, and start it at the top.

tooth :: (h :> a) -> Int

Return the index into the current Traversal within the current level of the zipper.

jerkTo (tooth l) l = Just'

Mnemonically, zippers have a number of teeth within each level. This is which tooth you are currently at.

up :: ((h :> s) :> a) -> h :> s

Move the zipper up, closing the current level and focusing on the parent element.

NB: Attempts to move up from the Top of the zipper will fail to typecheck.

>>> :t zipper ("hello","world") & down _1 & fromWithin traverse & up
zipper ("hello","world") & down _1 & fromWithin traverse & up
  :: (Top :> ([Char], [Char])) :> [Char]

right :: MonadPlus m => (h :> a) -> m (h :> a)

Jerk the zipper one tooth to the right within the current Lens or Traversal.

Attempts to move past the start of the current Traversal (or trivially, the current Lens) will return Nothing.

>>> isNothing $ zipper "hello" & right
True

>>> zipper "hello" & fromWithin traverse & right <&> view focus
'e'

>>> zipper "hello" & fromWithin traverse & right <&> focus .~ 'u' <&> rezip
"hullo"

>>> rezip $ zipper (1,2) & fromWithin both & tug right & focus .~ 3
(1,3)

left :: MonadPlus m => (h :> a) -> m (h :> a)

Jerk the zipper left one tooth within the current Lens or Traversal.

Attempts to move past the end of the current Traversal (or trivially, the current Lens) will return Nothing.

>>> isNothing $ zipper "hello" & left
True

tug :: (a -> Maybe a) -> a -> a

This allows you to safely 'tug left' or 'tug right' on a zipper. This will attempt the move, and stay where it was if it fails.

The more general signature allows its use in other circumstances, however.

tug f x ≡ fromMaybe a (f a)

>>> fmap rezip $ zipper "hello" & within traverse <&> tug left <&> focus .~ 'j'
"jello"

>>> fmap rezip $ zipper "hello" & within traverse <&> tug right <&> focus .~ 'u'
"hullo"

tugs :: (a -> Maybe a) -> Int -> a -> a

This allows you to safely tug left or tug right multiple times on a zipper, moving multiple steps in a given direction and stopping at the last place you couldn't move from. This lets you safely move a zipper, because it will stop at either end.

>>> fmap rezip $ zipper "stale" & within traverse <&> tugs right 2 <&> focus .~ 'y'
"style"

>>> rezip $ zipper "want" & fromWithin traverse & tugs right 2 & focus .~ 'r' & tugs left 100 & focus .~ 'c'
"cart"

farthest :: (a -> Maybe a) -> a -> a

Move in a direction as far as you can go, then stop there.

This repeatedly applies a function until it returns Nothing, and then returns the last answer.

>>> fmap rezip $ zipper ("hello","world") & down _1 & within traverse <&> farthest right <&> focus .~ 'a'
("hella","world")

>>> rezip $ zipper ("hello","there") & fromWithin (both.traverse) & farthest right & focus .~ 'm'
("hello","therm")

jerks :: Monad m => (a -> m a) -> Int -> a -> m a

This allows for you to repeatedly pull a zipper in a given direction, failing if it falls off the end.

>>> isNothing $ zipper "hello" & within traverse >>= jerks right 10
True

>>> fmap rezip $ zipper "silly" & within traverse >>= jerks right 3 <&> focus .~ 'k'
"silky"

teeth :: (h :> a) -> Int

Returns the number of siblings at the current level in the zipper.

teeth z >= 1

NB: If the current Traversal targets an infinite number of elements then this may not terminate.

>>> zipper ("hello","world") & teeth
1

>>> zipper ("hello","world") & fromWithin both & teeth
2

>>> zipper ("hello","world") & down _1 & teeth
1

>>> zipper ("hello","world") & down _1 & fromWithin traverse & teeth
5

>>> zipper ("hello","world") & fromWithin (_1.traverse) & teeth
5

>>> zipper ("hello","world") & fromWithin (both.traverse) & teeth
10

jerkTo :: MonadPlus m => Int -> (h :> a) -> m (h :> a)

Move the zipper horizontally to the element in the nth position in the current level, absolutely indexed, starting with the farthest left as 0.

This returns Nothing if the target element doesn't exist.

jerkTo n ≡ jerks right n . farthest left

>>> isNothing $ zipper "not working." & jerkTo 20
True

tugTo :: Int -> (h :> a) -> h :> a

Move the zipper horizontally to the element in the nth position of the current level, absolutely indexed, starting with the farthest left as 0.

If the element at that position doesn't exist, then this will clamp to the range 0 <= n < teeth.

tugTo n ≡ tugs right n . farthest left

>>> rezip $ zipper "not working." & fromWithin traverse & tugTo 100 & focus .~ '!' & tugTo 1 & focus .~ 'u'
"nut working!"

down :: SimpleLensLike (Context a a) s a -> (h :> s) -> (h :> s) :> a

Step down into a Lens. This is a constrained form of fromWithin for when you know there is precisely one target that can never fail.

 down :: Simple Lens s a -> (h :> s) -> h :> s :> a
 down :: Simple Iso s a  -> (h :> s) -> h :> s :> a

within :: MonadPlus m => SimpleLensLike (Bazaar a a) s a -> (h :> s) -> m ((h :> s) :> a)

Step down into the leftmost entry of a Traversal.

 within :: Simple Traversal s a -> (h :> s) -> Maybe (h :> s :> a)
 within :: Simple Lens s a      -> (h :> s) -> Maybe (h :> s :> a)
 within :: Simple Iso s a       -> (h :> s) -> Maybe (h :> s :> a)

withins :: SimpleLensLike (Bazaar a a) s a -> (h :> s) -> [(h :> s) :> a]

Step down into every entry of a Traversal simultaneously.

withins :: Simple Traversal s a -> (h :> s) -> [h :> s :> a] withins :: Simple Lens s a -> (h :> s) -> [h :> s :> a] withins :: Simple Iso s a -> (h :> s) -> [h :> s :> a]

fromWithin :: SimpleLensLike (Bazaar a a) s a -> (h :> s) -> (h :> s) :> a

Unsafely step down into a Traversal that is assumed to be non-empty.

If this invariant is not met then this will usually result in an error!

 fromWithin :: Simple Traversal s a -> (h :> s) -> h :> s :> a
 fromWithin :: Simple Lens s a      -> (h :> s) -> h :> s :> a
 fromWithin :: Simple Iso s a       -> (h :> s) -> h :> s :> a

You can reason about this function as if the definition was:

fromWithin l ≡ fromJust . within l

but it is lazier in such a way that if this invariant is violated, some code can still succeed if it is lazy enough in the use of the focused value.

class Zipper h a where

This enables us to pull the zipper back up to the Top.

Methods

recoil :: Coil h a -> [a] -> Zipped h a

rezip :: Zipper h a => (h :> a) -> Zipped h a

Close something back up that you opened as a zipper.

Tapes

data Tape k where

A Tape is a recorded path through the Traversal chain of a zipper.

Constructors

Tape :: Track h a -> !Int -> Tape (h :> a)

saveTape :: (h :> a) -> Tape (h :> a)

Save the current path as as a Tape we can play back later.

restoreTape :: MonadPlus m => Tape (h :> a) -> Zipped h a -> m (h :> a)

Restore ourselves to a previously recorded position precisely.

If the position does not exist, then fail.

restoreNearTape :: MonadPlus m => Tape (h :> a) -> Zipped h a -> m (h :> a)

Restore ourselves to a location near our previously recorded position.

When moving left to right through a Traversal, if this will clamp at each level to the range 0 <= k < teeth, so the only failures will occur when one of the sequence of downward traversals find no targets.

unsafelyRestoreTape :: Tape (h :> a) -> Zipped h a -> h :> a

Restore ourselves to a previously recorded position.

This *assumes* that nothing has been done in the meantime to affect the existence of anything on the entire path.

Motions left or right are clamped, but all traversals included on the Tape are assumed to be non-empty.

Violate these assumptions at your own risk!

Tracks

peel :: Coil h a -> Track h a

This is used to peel off the path information from a Coil for use when saving the current path for later replay.

data Track where

The Track forms the bulk of a Tape.

Constructors

Track :: Track Top a

Fork :: Track h s -> !Int -> SimpleLensLike (Bazaar a a) s a -> Track (h :> s) a

restoreTrack :: MonadPlus m => Track h a -> Zipped h a -> m (h :> a)

Restore ourselves to a previously recorded position precisely.

If the position does not exist, then fail.

restoreNearTrack :: MonadPlus m => Track h a -> Zipped h a -> m (h :> a)

Restore ourselves to a location near our previously recorded position.

When moving left to right through a Traversal, if this will clamp at each level to the range 0 <= k < teeth, so the only failures will occur when one of the sequence of downward traversals find no targets.

unsafelyRestoreTrack :: Track h a -> Zipped h a -> h :> a

Restore ourselves to a previously recorded position.

This *assumes* that nothing has been done in the meantime to affect the existence of anything on the entire path.

Motions left or right are clamped, but all traversals included on the Tape are assumed to be non-empty.

Violate these assumptions at your own risk!

Helper functions

reverseList :: [a] -> [a]

Reverse a list.

GHC doesn't optimize reverse [] into [], so we'll nudge it with our own reverse function.

This is relevant when descending into a lens, for example -- we know the unzipped part of the level will be empty.