Control.Lens.Internal.Zoom

Description

Zoom

newtype Focusing m s a

Used by Zoom to zoom into StateT.

Constructors

Focusing

Fields

unfocusing :: m (s, a)

newtype FocusingWith w m s a

Used by Zoom to zoom into RWST.

Constructors

FocusingWith

Fields

unfocusingWith :: m (s, a, w)

newtype FocusingPlus w k s a

Used by Zoom to zoom into WriterT.

Constructors

FocusingPlus

Fields

unfocusingPlus :: k (s, w) a

newtype FocusingOn f k s a

Used by Zoom to zoom into MaybeT or ListT.

Constructors

FocusingOn

Fields

unfocusingOn :: k (f s) a

newtype FocusingMay k s a

Used by Zoom to zoom into ErrorT.

Constructors

FocusingMay

Fields

unfocusingMay :: k (May s) a

newtype May a

Make a Monoid out of Maybe for error handling.

Constructors

May

Fields

getMay :: Maybe a

newtype FocusingErr e k s a

Used by Zoom to zoom into ErrorT.

Constructors

FocusingErr

Fields

unfocusingErr :: k (Err e s) a

newtype Err e a

Make a Monoid out of Either for error handling.

Constructors

Err

Fields

getErr :: Either e a

newtype FocusingFree f m k s a

Used by Zoom to zoom into FreeT

Constructors

FocusingFree

Fields

unfocusingFree :: k (Freed f m s) a

newtype Freed f m a

Make a Monoid out of FreeF for result collection.

Constructors

Freed

Fields

getFreed :: FreeF f a (FreeT f m a)

Magnify

newtype Effect m r a

Wrap a monadic effect with a phantom type argument.

Constructors

Effect

Fields

getEffect :: m r

newtype EffectRWS w st m s a

Wrap a monadic effect with a phantom type argument. Used when magnifying RWST.

Constructors

EffectRWS

Fields

getEffectRWS :: st -> m (s, st, w)