Control.Lens.Plated

Description

The name "plate" stems originally from "boilerplate", which was the term used by the "Scrap Your Boilerplate" papers, and later inherited by Neil Mitchell's "Uniplate".

http://community.haskell.org/~ndm/uniplate/

The combinators in here are designed to be compatible with and subsume the uniplate API with the notion of a Traversal replacing a uniplate or biplate.

By implementing these combinators in terms of plates instead of uniplate additional type safety is gained, as the user is no longer responsible for maintaining invariants such as the number of children he received.

Documentation

class Plated t where

A Plated type is one where we know how to extract its immediate self-similar children.

Example 1:

 data Expr
   = Val Int
   | Neg Expr
   | Add Expr Expr
   deriving (Eq,Ord,Show,Read)

 instance Plated Expr where
   plates f (Neg e) = Neg <$> f e
   plates f (Add a b) = Add <$> f a <*> f b
   plates _ a = pure a

Example 2:

 data Tree a
   = Bin (Tree a) (Tree a)
   | Tip a
   deriving (Eq,Ord,Show,Read)

 instance Plated (Tree a) where
   plates f (Bin l r) = Bin <$> f l <*> f r
   plates _ t = pure t

Methods

plates :: Simple Traversal t t

Traversal of the immediate children of this structure.

uniplate :: Plated a => a -> ([a], [a] -> a)

Neil Mitchell's uniplate combinator, implemented in terms of Plated.

uniplate = plateOf plates

plateOf :: LensLike (Kleene c c) a b c c -> a -> ([c], [c] -> b)

plateOf turns a Traversal into a uniplate (or biplate).

The resulting plate is actually safer to use as it replaces ignores 'over-application' and deals gracefully with under-application.

 plateOf :: Simple Iso a b       -> a -> ([b], [b] -> a)
 plateOf :: Simple Lens a b      -> a -> ([b], [b] -> a)
 plateOf :: Simple Traversal a b -> a -> ([b], [b] -> a)

children :: Plated a => a -> [a]

Extract the immediate descendants of a Plated container.

rewrite :: Plated a => (a -> Maybe a) -> a -> a

Rewrite by applying a rule everywhere you can. Ensures that the rule cannot be applied anywhere in the result:

propRewrite r x = all (isNothing . r) (universe (rewrite r x))

Usually transform is more appropriate, but rewrite can give better compositionality. Given two single transformations f and g, you can construct a -> f a mplus g a which performs both rewrites until a fixed point.

rewriteOf :: SimpleSetting a a -> (a -> Maybe a) -> a -> a

Rewrite by applying a rule everywhere you can. Ensures that the rule cannot be applied anywhere in the result:

propRewriteOf l r x = all (isNothing . r) (universeOf l (rewriteOf l r x))

Usually transformOf is more appropriate, but rewriteOf can give better compositionality. Given two single transformations f and g, you can construct a -> f a mplus g a which performs both rewrites until a fixed point.

 rewriteOf :: Simple Iso a a       -> (a -> Maybe a) -> a -> a
 rewriteOf :: Simple Lens a a      -> (a -> Maybe a) -> a -> a
 rewriteOf :: Simple Traversal a a -> (a -> Maybe a) -> a -> a
 rewriteOf :: Simple Setter a a    -> (a -> Maybe a) -> a -> a

rewriteOn :: Plated c => Setting a b c c -> (c -> Maybe c) -> a -> b

Rewrite recursively over part of a larger structure.

 rewriteOn :: Plated c => Simple Iso a b       -> (b -> Maybe b) -> a -> a
 rewriteOn :: Plated c => Simple Lens a b      -> (b -> Maybe b) -> a -> a
 rewriteOn :: Plated c => Simple Traversal a b -> (b -> Maybe b) -> a -> a
 rewriteOn :: Plated c => Simple Setting a b   -> (b -> Maybe b) -> a -> a

rewriteOnOf :: Setting a b c c -> SimpleSetting c c -> (c -> Maybe c) -> a -> b

Rewrite recursively over part of a larger structure using a specified setter.

 rewriteOnOf :: Plated b => Simple Iso a b       -> Simple Iso b b       -> (b -> Maybe b) -> a -> a
 rewriteOnOf :: Plated b => Simple Lens a b      -> Simple Lens b b      -> (b -> Maybe b) -> a -> a
 rewriteOnOf :: Plated b => Simple Traversal a b -> Simple Traversal b b -> (b -> Maybe b) -> a -> a
 rewriteOnOf :: Plated b => Simple Setter a b    -> Simple Setter b b    -> (b -> Maybe b) -> a -> a

rewriteM :: (Monad m, Plated a) => (a -> m (Maybe a)) -> a -> m a

Rewrite by applying a monadic rule everywhere you can. Ensures that the rule cannot be applied anywhere in the result.

rewriteMOf :: Monad m => SimpleLensLike (WrappedMonad m) a a -> (a -> m (Maybe a)) -> a -> m a

Rewrite by applying a monadic rule everywhere you recursing with a user-specified Traversal. Ensures that the rule cannot be applied anywhere in the result.

rewriteMOn :: (Monad m, Plated c) => LensLike (WrappedMonad m) a b c c -> (c -> m (Maybe c)) -> a -> m b

Rewrite by applying a monadic rule everywhere inside of a structure located by a user-specified Traversal. Ensures that the rule cannot be applied anywhere in the result.

rewriteMOnOf :: Monad m => LensLike (WrappedMonad m) a b c c -> SimpleLensLike (WrappedMonad m) c c -> (c -> m (Maybe c)) -> a -> m b

Rewrite by applying a monadic rule everywhere inside of a structure located by a user-specified Traversal, using a user-specified Traversal for recursion. Ensures that the rule cannot be applied anywhere in the result.

universe :: Plated a => a -> [a]

Get all the descendants of a node, including itself.

 universe (Add (Val 1) (Neg (Val 2))) = [Add (Val 1) (Neg (Val 2)), Val 1, Neg (Val 2), Val 2]

This method is often combined with a list comprehension, for example:

vals x = [i | Val i <- universe x]

tails = universe

universeOf :: Fold a a -> a -> [a]

Given a traversal that knows how to locate immediate children, retrieve all of the transitive descendants of a node, including itself.

universeOn :: Plated b => Fold a b -> a -> [b]

Get all of the descendants of a node that lie in a region indicated by a Fold (or Getter)

 universeOn :: Plated b => Getter a b           -> a -> [b]
 universeOn :: Plated b => Fold a b             -> a -> [b]
 universeOn :: Plated b => Simple Lens a b      -> a -> [b]
 universeOn :: Plated b => Simple Traversal a b -> a -> [b]
 universeOn :: Plated b => Simple Iso a b       -> a -> [b]

universeOnOf :: Fold a b -> Fold b b -> a -> [b]

Given a traversal that knows how to locate immediate children, retrieve all of the transitive descendants of a node, including itself that lie in a region indicated by another fold.

transform :: Plated a => (a -> a) -> a -> a

Transform every element in the tree, in a bottom-up manner.

For example, replacing negative literals with literals:

 negLits = transform $ x -> case x of
   Neg (Lit i) -> Lit (negate i)
   _           -> x

transformOf :: SimpleSetting a a -> (a -> a) -> a -> a

Transform every element by recursively applying a given Setter in a bottom-up manner.

 transformOf :: Simple Traversal a a -> (a -> a) -> a -> a
 transformOf :: Simple Setter a a    -> (a -> a) -> a -> a

transformOn :: Plated c => Setting a b c c -> (c -> c) -> a -> b

Transform every element in the tree in a bottom-up manner over a region indicated by a Setter.

 transformOn :: Plated b => Simple Traversal a b -> (b -> b) -> a -> a
 transformOn :: Plated b => Simple Setter a b    -> (b -> b) -> a -> a

transformOnOf :: Setting a b c c -> SimpleSetting c c -> (c -> c) -> a -> b

Transform every element in a region indicated by a Setter by recursively applying another Setter in a bottom-up manner.

 transformOnOf :: Setter a b -> Simple Traversal b b -> (b -> b) -> a -> a
 transformOnOf :: Setter a b -> Simple Setter b b    -> (b -> b) -> a -> a

transformM :: (Monad m, Plated a) => (a -> m a) -> a -> m a

Transform every element in the tree, in a bottom-up manner, monadically.

transformMOf :: Monad m => SimpleLensLike (WrappedMonad m) a a -> (a -> m a) -> a -> m a

Transform every element in a tree using a user supplied Traversal in a bottom-up manner with a monadic effect.

 transformMOf :: Monad m => 'Simple Traversal a a -> (a -> m a) -> a -> m a

transformMOn :: (Monad m, Plated c) => LensLike (WrappedMonad m) a b c c -> (c -> m c) -> a -> m b

Transform every element in the tree in a region indicated by a supplied Traversal, in a bottom-up manner, monadically.

 transformMOn :: (Monad m, Plated c) => Simple Traversal a b -> (b -> m b) -> a -> m a

transformMOnOf :: Monad m => LensLike (WrappedMonad m) a b c c -> SimpleLensLike (WrappedMonad m) c c -> (c -> m c) -> a -> m b

Transform every element in a tree that lies in a region indicated by a supplied Traversal, walking with a user supplied Traversal in a bottom-up manner with a monadic effect.

 transformMOnOf :: Monad m => Simple Traversal a b -> Simple Traversal b b -> (b -> m b) -> a -> m a

descend :: Plated a => (a -> a) -> a -> a

Recurse one level into a structure.

descend = over plates

descendOf :: Setting a b c d -> (c -> d) -> a -> b

Recurse one level into a structure using a user specified recursion scheme. This is over, but it is supplied here for consistency with the uniplate API.

descendOf = over

 descendOf :: Simple Setter a b -> (b -> b) -> a -> a
 descendOf :: Simple Traversal a b -> (b -> b) -> a -> a

descendOn :: Plated c => Setting a b c c -> (c -> c) -> a -> b

Recurse one level into the parts of the structure delimited by a Setter.

descendOn b = over (b . plates)

 descendOn :: Plated c => Setter a b -> (b -> b) -> a -> a

descendOnOf :: Setting a b c d -> Setting c d e f -> (e -> f) -> a -> b

Recurse one level into the parts delimited by one Setter, using another.

descendOnOf b l = over (b.l)

 descendOnOf :: Simple Setter a b    -> Simple Setter b b    -> (b -> b) -> a -> a
 descendOnOf :: Simple Traversal a b -> Simple Traversal b b -> (b -> b) -> a -> a

descendA :: (Applicative f, Plated a) => (a -> f a) -> a -> f a

Recurse one level into a structure with an Applicative effect, this is plates, but it is supplied for consistency with the uniplate API.

descendA = plates

descendAOf :: Applicative f => LensLike f a b c d -> (c -> f d) -> a -> f b

Recurse one level into a structure using a user specified recursion scheme and Applicative effects. This is id, but it is supplied for consistency with the uniplate API.

descendAOf = id

 descendAOf :: Applicative m => Simple Traversal a b => (b -> m b) -> a -> m a

descendAOn :: (Applicative f, Plated c) => LensLike f a b c c -> (c -> f c) -> a -> f b

Recurse one level into the parts of the structure delimited by a Traversal with Applicative effects.

descendAOn b = b . plates

 descendAOn :: (Applicative f, Plated' c) => Simple Traversal a b -> (b -> f b) -> a -> f a

descendAOnOf :: Applicative g => LensLike g a b c d -> LensLike g c d e f -> (e -> g f) -> a -> g b

Recurse one level into the parts delimited by one Traversal, using another with Applicative effects.

descendAOnOf = (.)

 descendAOnOf :: Applicative f => Simple Traversal a b -> Simple Traversal b b -> (b -> f b) -> a -> f a

descendM :: (Monad m, Plated a) => (a -> m a) -> a -> m a

Recurse one level into a structure with a monadic effect.

descendM = mapMOf plates

descendMOf :: Monad m => LensLike (WrappedMonad m) a b c d -> (c -> m d) -> a -> m b

Recurse one level into a structure using a user specified recursion scheme and monadic effects. This is id, but it is supplied for consistency with the uniplate API.

descendMOf = mapMOf

 descendMOf :: Monad m => Simple Traversal a b => (b -> m b) -> a -> m a

descendMOn :: (Monad m, Plated c) => LensLike (WrappedMonad m) a b c c -> (c -> m c) -> a -> m b

Recurse one level into the parts of the structure delimited by a Traversal with monadic effects.

descendMOn b = mapMOf (b . plates)

 descendMOn :: (Monad f, Plated c) => Simple Traversal a b -> (b -> f b) -> a -> f a

descendMOnOf :: Monad m => LensLike (WrappedMonad m) a b c c -> SimpleLensLike (WrappedMonad m) c c -> (c -> m c) -> a -> m b

Recurse one level into the parts delimited by one Traversal, using another with monadic effects. |

descendMOnOf b l = mapMOf (b . l)

 descendMOnOf :: Monad m => Simple Traversal a b -> Simple Traversal b b -> (b -> m b) -> a -> m a

contexts :: Plated a => a -> [Context a a a]

Return a list of all of the editable contexts for every location in the structure, recursively.

 propUniverse x = universe x == map pos (contexts x)
 propId x = all (== x) [extract w | w <- contexts x]

contexts = contextsOf plates

contextsOf :: SimpleLensLike (Kleene a a) a a -> a -> [Context a a a]

Return a list of all of the editable contexts for every location in the structure, recursively, using a user-specified Traversal to walk each layer.

 propUniverse l x = universeOf l x == map pos (contextsOf l x)
 propId l x = all (== x) [extract w | w <- contextsOf l x]

 contextsOf :: Simple Traversal a a -> a -> [Context a a]

contextsOn :: Plated c => LensLike (Kleene c c) a b c c -> a -> [Context c c b]

Return a list of all of the editable contexts for every location in the structure in an areas indicated by a user supplied Traversal, recursively using plates.

contextsOn b = contextsOnOf b plates

 contextsOn :: Plated b => Simple Traversal a b -> a -> [Context b b a]

contextsOnOf :: LensLike (Kleene c c) a b c c -> SimpleLensLike (Kleene c c) c c -> a -> [Context c c b]

Return a list of all of the editable contexts for every location in the structure in an areas indicated by a user supplied Traversal, recursively using another user-supplied Traversal to walk each layer.

 contextsOnOf :: Simple Traversal a b -> Simple Traversal b b -> a -> [Context b b a]

holes :: Plated a => a -> [Context a a a]

The one-level version of context. This extracts a list of the immediate children as editable contexts.

Given a context you can use pos to see the values, peek at what the structure would be like with an edited result, or simply extract the original structure.

 propChildren x = children l x == map pos (holes l x)
 propId x = all (== x) [extract w | w <- holes l x]

holes = holesOf plates

holesOf :: LensLike (Kleene c c) a b c c -> a -> [Context c c b]

The one-level version of contextsOf. This extracts a list of the immediate children according to a given Traversal as editable contexts.

Given a context you can use pos to see the values, peek at what the structure would be like with an edited result, or simply extract the original structure.

 propChildren l x = childrenOf l x == map pos (holesOf l x)
 propId l x = all (== x) [extract w | w <- holesOf l x]

 holesOf :: Simple Iso a b       -> a -> [Context b a]
 holesOf :: Simple Lens a b      -> a -> [Context b a]
 holesOf :: Simple Traversal a b -> a -> [Context b a]

holesOn :: LensLike (Kleene c c) a b c c -> a -> [Context c c b]

An alias for holesOf, provided for consistency with the other combinators.

holesOn = holesOf

 holesOn :: Simple Iso a b       -> a -> [Context b b a]
 holesOn :: Simple Lens a b      -> a -> [Context b b a]
 holesOn :: Simple Traversal a b -> a -> [Context b b a]

holesOnOf :: LensLike (Kleene e e) a b c d -> LensLike (Kleene e e) c d e e -> a -> [Context e e b]

Extract one level of holes from a container in a region specified by one Traversal, using another.

holesOnOf b l = holesOf (b.l)

 holesOnOf :: 'Simple Iso a b       -> Simple Iso b b       -> a -> [Context b a]
 holesOnOf :: 'Simple Lens a b      -> Simple Lens b b      -> a -> [Context b a]
 holesOnOf :: 'Simple Traversal a b -> Simple Traversal b b -> a -> [Context b a]

Operations to be careful of

unsafePlateOf :: LensLike (Kleene c d) a b c d -> a -> ([c], [d] -> b)

unsafePlateOf turns a Traversal into a uniplate (or biplate) family.

If you do not need the types of c and d to be different, it is recommended that you use plateOf

It is generally safer to traverse with the Kleene indexed store Comonad rather than use this combinator. However, it is sometimes convenient.

This is unsafe because if you don't supply at least as many d's as you were given c's, then the reconstruction of b will result in an error.