Control.Lens.Prism

Description

Prisms

type Prism s t a b = forall p f. (Prismatic p, Applicative f) => p a (f b) -> p s (f t)

A Prism l is a 0-or-1 target Traversal that can also be turned around with remit to obtain a Getter in the opposite direction.

There are two laws that a Prism should satisfy:

First, if I remit or review a value with a Prism and then preview or use (^?), I will get it back:

preview l (review l b) ≡ Just b

Second, if you can extract a value a using a Prism l from a value s, then the value s is completely described my l and a:

If preview l s ≡ Just a then review l a ≡ s

These two laws imply that the Traversal laws hold for every Prism and that we traverse at most 1 element:

lengthOf l x <= 1

It may help to think of this as a Iso that can be partial in one direction.

Every Prism is a valid Traversal.

Every Iso is a valid Prism.

For example, you might have a Prism' Integer Natural allows you to always go from a Natural to an Integer, and provide you with tools to check if an Integer is a Natural and/or to edit one if it is.

 nat :: Prism' Integer Natural
 nat = prism toInteger $ \ i ->
    if i < 0
    then Left i
    else Right (fromInteger i)

Now we can ask if an Integer is a Natural.

>>> 5^?nat
Just 5

>>> (-5)^?nat
Nothing

We can update the ones that are:

>>> (-3,4) & both.nat *~ 2
(-3,8)

And we can then convert from a Natural to an Integer.

>>> 5 ^. remit nat -- :: Natural
5

Similarly we can use a Prism to traverse the left half of an Either:

>>> Left "hello" & _left %~ length
Left 5

or to construct an Either:

>>> 5^.remit _left
Left 5

such that if you query it with the Prism, you will get your original input back.

>>> 5^.remit _left ^? _left
Just 5

Another interesting way to think of a Prism is as the categorical dual of a Lens -- a co-Lens, so to speak. This is what permits the construction of outside.

type Prism' s a = Prism s s a a

A Simple Prism

type APrism s t a b = Market a b a (Mutator b) -> Market a b s (Mutator t)

If you see this in a signature for a function, the function is expecting a Prism.

type APrism' s a = APrism s s a a

type Reviewing s t a b = Overloading Review Review Identity s t a b

If you see this in a signature for a function, the function is expecting a Prism

type Reviewing' s a = Reviewing s s a a

Constructing Prisms

prism :: (b -> t) -> (s -> Either t a) -> Prism s t a b

Build a Prism.

Either t a is used instead of Maybe a to permit the types of s and t to differ.

prism' :: (a -> s) -> (s -> Maybe a) -> Prism' s a

Build a Prism'.

Consuming Prisms

clonePrism :: APrism s t a b -> Prism s t a b

Clone a Prism so that you can reuse the same monomorphically typed Prism for different purposes.

See cloneLens and cloneTraversal for examples of why you might want to do this.

withPrism :: ((b -> t) -> (s -> Either t a) -> r) -> APrism s t a b -> r

Safely decompose APrism

remit :: Reviewing s t a b -> Getter b t

Turn a Prism or Iso around to build a Getter.

If you have an Iso, from is a more powerful version of this function that will return an Iso instead of a mere Getter.

>>> 5 ^.remit _left
Left 5

 remit :: Prism s t a b -> Getter b t
 remit :: Iso s t a b   -> Getter b t

review :: MonadReader b m => Reviewing s t a b -> m t

This can be used to turn an Iso or Prism around and view a value (or the current environment) through it the other way.

review ≡ view . remit

>>> review _left "mustard"
Left "mustard"

Usually review is used in the (->) monad with a Prism' or Iso, in which case it may be useful to think of it as having one of these more restricted type signatures:

 review :: Iso' s a        -> a -> s
 review :: Prism' s a -> a -> s

However, when working with a monad transformer stack, it is sometimes useful to be able to review the current environment, in which case one of these more slightly more liberal type signatures may be beneficial to think of it as having:

 review :: MonadReader a m => Iso' s a        -> m s
 review :: MonadReader a m => Prism' s a -> m s

reviews :: MonadReader b m => Reviewing s t a b -> (t -> r) -> m r

This can be used to turn an Iso or Prism around and view a value (or the current environment) through it the other way, applying a function.

reviews ≡ views . remit

>>> reviews _left isRight "mustard"
False

Usually this function is used in the (->) monad with a Prism' or Iso, in which case it may be useful to think of it as having one of these more restricted type signatures:

 reviews :: Iso' s a        -> (s -> r) -> a -> r
 reviews :: Prism' s a -> (s -> r) -> a -> r

However, when working with a monad transformer stack, it is sometimes useful to be able to review the current environment, in which case one of these more slightly more liberal type signatures may be beneficial to think of it as having:

 reviews :: MonadReader a m => Iso' s a   -> (s -> r) -> m r
 reviews :: MonadReader a m => Prism' s a -> (s -> r) -> m r

reuse :: MonadState b m => Reviewing s t a b -> m t

This can be used to turn an Iso or Prism around and use a value (or the current environment) through it the other way.

reuse ≡ use . remit

>>> evalState (reuse _left) 5
Left 5

 reuse :: MonadState a m => Prism' s a -> m s
 reuse :: MonadState a m => Iso' s a   -> m s

reuses :: MonadState b m => Reviewing s t a b -> (t -> r) -> m r

This can be used to turn an Iso or Prism around and use the current state through it the other way, applying a function.

reuses ≡ uses . remit

>>> evalState (reuses _left isLeft) (5 :: Int)
True

 reuses :: MonadState a m => Prism' s a -> (s -> r) -> m r
 reuses :: MonadState a m => Iso' s a   -> (s -> r) -> m r

outside :: APrism s t a b -> Lens (t -> r) (s -> r) (b -> r) (a -> r)

Use a Prism as a kind of first-class pattern.

outside :: Prism s t a b -> Lens (t -> r) (s -> r) (b -> r) (a -> r)

aside :: APrism s t a b -> Prism (e, s) (e, t) (e, a) (e, b)

Use a Prism to work over part of a structure.

without :: APrism s t a b -> APrism u v c d -> Prism (Either s u) (Either t v) (Either a c) (Either b d)

Given a pair of prisms, project sums.

Viewing a Prism as a co-lens, this combinator can be seen to be dual to alongside.

Common Prisms

_left :: Prism (Either a c) (Either b c) a b

This prism provides a traversal for tweaking the left-hand value of an Either:

>>> over _left (+1) (Left 2)
Left 3

>>> over _left (+1) (Right 2)
Right 2

>>> Right 42 ^._left :: String
""

>>> Left "hello" ^._left
"hello"

It also can be turned around to obtain the embedding into the Left half of an Either:

>>> 5^.remit _left
Left 5

_right :: Prism (Either c a) (Either c b) a b

This prism provides a traversal for tweaking the right-hand value of an Either:

>>> over _right (+1) (Left 2)
Left 2

>>> over _right (+1) (Right 2)
Right 3

>>> Right "hello" ^._right
"hello"

>>> Left "hello" ^._right :: [Double]
[]

It also can be turned around to obtain the embedding into the Right half of an Either:

>>> 5^.remit _right
Right 5

_just :: Prism (Maybe a) (Maybe b) a b

This prism provides a traversal for tweaking the target of the value of Just in a Maybe.

>>> over _just (+1) (Just 2)
Just 3

Unlike traverse this is a Prism, and so you can use it to inject as well:

>>> 5^.remit _just
Just 5

Implementation details

class Profunctor p => Prismatic p where

Methods

prismatic :: Applicative f => (s -> Either t a) -> p a (f t) -> p s (f t)