Control.Lens.Projection

Description

Projections

type Projection s t a b = forall k f. (Projective k, Applicative f) => k (a -> f b) (s -> f t)

A Projection is a Traversal that can also be turned around with qua to obtain a Getter

Constructing Projections

class Isomorphic k => Projective k where

Used to provide overloading of projections.

An instance of Projective is a Category with a canonical mapping to it from the category of embedding-projection pairs over Haskell types.

Methods

projecting :: Applicative f => (b -> t) -> ((a -> f b) -> s -> f t) -> k (a -> f b) (s -> f t)

Build a Projection or Traversal

data Project x y where

This data type is used to capture all of the information provided by the Projective class, so you can turn a Projection around into a Getter or otherwise muck around with its internals.

If you see a function that expects a Project, it is probably just expecting a Projection. You can safely

Constructors

Project :: (b -> t) -> ((a -> f b) -> s -> f t) -> Project (a -> f b) (s -> f t)

Consuming Projections

type Projecting f s t a b = Overloaded Project f s t a b

Consume a Project. This is commonly used when a function takes a Projection as a parameter.

project :: (Projective k, Applicative f) => Overloaded Project f s t a b -> Overloaded k f s t a b

Reflect a Projection.

qua :: Projecting Mutator s t a b -> Getter b t

Turn a Projection around to get at its contents.

review :: MonadReader b m => Projecting Mutator s t a b -> m t

This can be used to turn an Iso or Projection around and view a value (or the current environment) through it the other way.

review ≡ view . from

reviews :: MonadReader b m => Projecting Mutator s t a b -> (t -> r) -> m r

This can be used to turn an Iso or Projection around and view a value (or the current environment) through it the other way, applying a function.

reviews ≡ views . from

reuse :: MonadState b m => Projecting Mutator s t a b -> m t

This can be used to turn an Iso or Projection around and use a value (or the current environment) through it the other way.

reuse ≡ use . from

reuses :: MonadState b m => Projecting Mutator s t a b -> (t -> r) -> m r

This can be used to turn an Iso or Projection around and use the current state through it the other way, applying a function.

reuses ≡ uses . from

Common projections

_left :: Projection (Either a c) (Either b c) a b

A traversal for tweaking the left-hand value of an Either:

>>> over _left (+1) (Left 2)
Left 3

>>> over _left (+1) (Right 2)
Right 2

>>> Right 42 ^._left :: String
""

>>> Left "hello" ^._left
"hello"

_left :: Applicative f => (a -> f b) -> Either a c -> f (Either b c)

_right :: Projection (Either c a) (Either c b) a b

traverse the right-hand value of an Either:

_right ≡ traverse

Unfortunately the instance for Traversable (Either c) is still missing from base, so this can't just be traverse

>>> over _right (+1) (Left 2)
Left 2

>>> over _right (+1) (Right 2)
Right 3

>>> Right "hello" ^._right
"hello"

>>> Left "hello" ^._right :: [Double]
[]

_right :: Applicative f => (a -> f b) -> Either c a -> f (Either c a)

Simple

type SimpleProjection s a = Projection s s a a

A Simple Projection.