Control.Lens.Projection

Description

Documentation

type Projection s t a b = forall k f. (Projective k s b, Applicative f) => k (a -> f b) (s -> f s)

A Projection is a Traversal that can also be turned around with by to obtain a Getter

class Projective k a d where

Used to provide overloading of projections.

Methods

projective :: (d -> a) -> (x -> y) -> k x y

project :: Projective k s b => Overloaded (Project s b) f s s a b -> Overloaded k f s s a b

Reflect a Projection.

by :: Project s b (b -> Identity b) (s -> Identity s) -> Getter b s

Turn a Projection around to get an embedding

data Project s b x y

A concrete Projection, suitable for storing in a container or extracting an embedding.

Constructors

Project (b -> s) (x -> y)

projection :: (b -> s) -> (s -> Maybe a) -> Projection s t a b

Build a Projection

stereo :: Projective k s a => Project t a y z -> Project s t x y -> k x z

Compose projections.

mirror :: Projective k s a => Simple Iso s a -> Simple Projection s a

Convert an Iso to a Projection.

Ideally we would be able to use an Iso directly as a Projection, but this opens a can of worms.

Simple

type SimpleProjection s a = Projection s s a a

type SimpleProjection = Simple Projection