Control.Lens.Projection

Description

Documentation

type Projection a b c d = forall k f. (Projective k a d, Applicative f) => k (c -> f d) (a -> f a)

A Projection is a Traversal that can also be turned around with by to obtain a Getter

class Projective k a d where

Used to provide overloading of projections.

Methods

projective :: (d -> a) -> (x -> y) -> k x y

project :: Projective k a d => Overloaded (Project a d) f a a c d -> Overloaded k f a a c d

Reflect a Projection.

by :: Project a d (d -> Identity d) (a -> Identity a) -> Getter d a

Turn a Projection around to get an embedding

data Project a d x y

A concrete Projection, suitable for storing in a container or extracting an embedding.

Constructors

Project (d -> a) (x -> y)

projection :: (d -> a) -> (a -> Maybe c) -> Projection a b c d

Build a Projection

stereo :: Projective k a c => Project b c y z -> Project a b x y -> k x z

Compose projections.

mirror :: Projective k a c => Simple Iso a c -> Simple Projection a c

Convert an Iso to a Projection.

Ideally we would be able to use an Iso directly as a Projection, but this opens a can of worms.

Simple

type SimpleProjection a b = Projection a a b b

type SimpleProjection = Simple Projection