Control.Lens.Wrapped

Description

The Wrapped class provides similar functionality as Control.Newtype, from the newtype package, but in a more convenient and efficient form.

There are a few functions from newtype that are not provided here, because they can be done with the Iso directly:

 Control.Newtype.over Sum f ≡ wrapping Sum %~ f
 Control.Newtype.under Sum f ≡ unwrapping Sum %~ f
 Control.Newtype.overF Sum f ≡ mapping (wrapping Sum) %~ f
 Control.Newtype.underF Sum f ≡ mapping (unwrapping Sum) %~ f

under can also be used with wrapping to provide the equivalent of Control.Newtype.under. Also, most use cases don't need full polymorphism, so only the single constructor wrapping functions would be needed.

These equivalences aren't 100% honest, because newtype's operators need to rely on two Newtype constraints. This means that the wrapper used for the output is not necessarily the same as the input.

Documentation

class Wrapped s t a b | a -> s, b -> t, a t -> s, b s -> t where

Wrapped provides isomorphisms to wrap and unwrap newtypes or data types with one constructor.

Methods

wrapped :: Iso s t a b

An isomorphism between s and a and a related one between t and b, such that when a = b, s = t.

This is often used via wrapping to aid type inference.

unwrapped :: Wrapped t s b a => Iso a b s t

This is a convenient alias for from wrapped.

>>> Const "hello" & unwrapped %~ length & getConst
5

wrapped' :: Wrapped s s a a => Iso' s a

A convenient type-restricted version of wrapped for aiding type inference.

unwrapped' :: Wrapped s s a a => Iso' a s

A convenient type-restricted version of unwrapped for aiding type inference.

wrapping :: Wrapped s s a a => (s -> a) -> Iso s s a a

This is a convenient version of wrapped with an argument that's ignored.

The argument is used to specify which newtype the user intends to wrap by using the constructor for that newtype.

The user supplied function is ignored, merely its type is used.

unwrapping :: Wrapped s s a a => (s -> a) -> Iso a a s s

This is a convenient version of unwrapped with an argument that's ignored.

The argument is used to specify which newtype the user intends to remove by using the constructor for that newtype.

The user supplied function is ignored, merely its type is used.

wrappings :: Wrapped s t a b => (s -> a) -> (t -> b) -> Iso s t a b

This is a convenient version of wrapped with two arguments that are ignored.

These arguments are used to which newtype the user intends to wrap and should both be the same constructor. This redundancy is necessary in order to find the full polymorphic isomorphism family.

The user supplied functions are ignored, merely their types are used.

unwrappings :: Wrapped t s b a => (s -> a) -> (t -> b) -> Iso a b s t

This is a convenient version of unwrapped with two arguments that are ignored.

These arguments are used to which newtype the user intends to remove and should both be the same constructor. This redundancy is necessary in order to find the full polymorphic isomorphism family.

The user supplied functions are ignored, merely their types are used.

op :: Wrapped s s a a => (s -> a) -> a -> s

Given the constructor for a Wrapped type, return a deconstructor that is its inverse.

Assuming the Wrapped instance is legal, these laws hold:

 op f . f ≡ id
 f . op f ≡ id

>>> op Identity (Identity 4)
4

>>> op Const (Const "hello")
"hello"

ala :: Wrapped s s a a => (s -> a) -> ((s -> a) -> e -> a) -> e -> s

This combinator is based on ala from Conor McBride's work on Epigram.

As with wrapping, the user supplied function for the newtype is ignored.

>>> ala Sum foldMap [1,2,3,4]
10

>>> ala All foldMap [True,True]
True

>>> ala All foldMap [True,False]
False

>>> ala Any foldMap [False,False]
False

>>> ala Any foldMap [True,False]
True

>>> ala Sum foldMap [1,2,3,4]
10

>>> ala Product foldMap [1,2,3,4]
24

alaf :: Wrapped s s a a => (s -> a) -> ((r -> a) -> e -> a) -> (r -> s) -> e -> s

This combinator is based on ala' from Conor McBride's work on Epigram.

As with wrapping, the user supplied function for the newtype is ignored.

>>> alaf Sum foldMap length ["hello","world"]
10