Control.Lens.Zoom

Description

Documentation

class (MonadReader b m, MonadReader a n) => Magnify m n k b a | m -> b, n -> a, m a -> n, n b -> m where

This class allows us to use magnify part of the environment, changing the environment supplied by many different monad transformers. Unlike zoom this can change the environment of a deeply nested monad transformer.

Also, unlike zoom, this can be used with any valid Getter, but cannot be used with a Traversal or Fold.

Methods

magnify :: ((b -> k c b) -> a -> k c a) -> m c -> n c

Run a monadic action in a larger environment than it was defined in, using a Getter.

This acts like local, but can in many cases change the type of the environment as well.

This is commonly used to lift actions in a simpler Reader monad into a monad with a larger environment type.

This can be used to edit pretty much any monad transformer stack with an environment in it:

 magnify ::             Getter s a -> (a -> r) -> s -> r
 magnify :: Monoid c => Fold s a   -> (a -> r) -> s -> r
 magnify :: Monoid w                Getter s t -> RWST s w st c -> RWST t w st c
 magnify :: (Monoid w, Monoid c) => Fold s t   -> RWST s w st c -> RWST t w st c
 ...

class (MonadState s m, MonadState t n) => Zoom m n k s t | m -> s k, n -> t k, m t -> n, n s -> m where

This class allows us to use zoom in, changing the State supplied by many different monad transformers, potentially quite deep in a monad transformer stack.

Methods

zoom :: Monad m => SimpleLensLike (k c) t s -> m c -> n c

Run a monadic action in a larger state than it was defined in, using a Simple Lens or Simple Traversal.

This is commonly used to lift actions in a simpler state monad into a state monad with a larger state type.

When applied to a 'Simple Traversal over multiple values, the actions for each target are executed sequentially and the results are aggregated.

This can be used to edit pretty much any monad transformer stack with a state in it!

 zoom :: Monad m             => Simple Lens s t      -> StateT t m a -> StateT s m a
 zoom :: (Monad m, Monoid c) => Simple Traversal s t -> StateT t m c -> StateT s m c
 zoom :: Monad m             => Simple Lens s t      -> RWST r w t m c -> RWST r w s m c
 zoom :: (Monad m, Monoid c) => Simple Traversal s t -> RWST r w t m c -> RWST r w s m c
 zoom :: Monad m             => Simple Lens s t      -> ErrorT e (RWST r w t m c) -> ErrorT e (RWST r w s m c)
 zoom :: (Monad m, Monoid c) => Simple Traversal s t -> ErrorT e (RWST r w t m c) -> ErrorT e (RWST r w s m c)
 ...