Control.Seq.Lens

Description

A Fold can be used to take the role of Foldable in Control.Seq.

Documentation

seqOf :: Getting (Endo [a]) s a -> Strategy a -> Strategy s

Evaluate the elements targeted by a Lens, Traversal, Iso, Getter or Fold according to the given strategy.

seqFoldable = seqOf folded