Data.Dynamic.Lens

Description

Documentation

traverseDynamic :: (Typeable a, Typeable b) => Traversal Dynamic Dynamic a b

Traverse the typed value contained in a Dynamic where the type required by your function matches that of the contents of the Dynamic.

 traverseDynamic :: (Applicative f, Typeable a, Typeable b) => (a -> f b) -> Dynamic -> f Dynamic