Data.Dynamic.Lens

Description

Documentation

dynamic :: Typeable a => Prism' Dynamic a

Traverse the typed value contained in a Dynamic where the type required by your function matches that of the contents of the Dynamic.