Data.Dynamic.Lens

Description

Documentation

class AsDynamic t where

Any Dynamic can be thrown as an Exception

Minimal complete definition

Nothing

Methods

_Dynamic :: Typeable a => Prism' t a

This Prism allows you to traverse the typed value contained in a Dynamic where the type required by your function matches that of the contents of the Dynamic, or construct a Dynamic value out of whole cloth. It can also be used to catch or throw a Dynamic value as SomeException.

_Dynamic :: Typeable a => Prism' Dynamic       a
_Dynamic :: Typeable a => Prism' SomeException a