Data.HashSet.Lens

Description

Documentation

setmapped :: (Eq j, Hashable j) => IndexPreservingSetter (HashSet i) (HashSet j) i j

This Setter can be used to change the type of a HashSet by mapping the elements to new values.

Sadly, you can't create a valid Traversal for a Set, but you can manipulate it by reading using folded and reindexing it via setmapped.

setOf :: Hashable a => Getting (HashSet a) s a -> s -> HashSet a

Construct a set from a Getter, Fold, Traversal, Lens or Iso.

setOf :: Hashable a         => Getter s a     -> s -> HashSet a
setOf :: (Eq a, Hashable a) => Fold s a       -> s -> HashSet a
setOf :: Hashable a         => Iso' s a       -> s -> HashSet a
setOf :: Hashable a         => Lens' s a      -> s -> HashSet a
setOf :: (Eq a, Hashable a) => Traversal' s a -> s -> HashSet a