Data.IntMap.Lens

Description

Documentation

traverseAtMin :: SimpleIndexedTraversal Int (IntMap v) v

Traverse the value at the minimum key in a Map

The key of the minimum element is available as the index.

traverseAtMax :: SimpleIndexedTraversal Int (IntMap v) v

Traverse the value at the maximum key in a Map