Data.List.Lens

Description

Traversals for manipulating parts of a list.

Documentation

prefixed :: Eq a => [a] -> Prism' [a] [a]

A Prism stripping a prefix from a list when used as a Traversal, or prepending that prefix when run backwards:

>>> "preview" ^? prefixed "pre"
Just "view"

>>> "review" ^? prefixed "pre"
Nothing

>>> prefixed "pre" # "amble"
"preamble"

suffixed :: Eq a => [a] -> Prism' [a] [a]

A Prism stripping a suffix from a list when used as a Traversal, or appending that suffix when run backwards:

>>> "review" ^? suffixed "view"
Just "re"

>>> "review" ^? suffixed "tire"
Nothing

>>> suffixed ".o" # "hello"
"hello.o"

stripSuffix :: Eq a => [a] -> [a] -> Maybe [a]