Data.Map.Lens

Description

Documentation

traverseAtMin :: SimpleIndexedTraversal k (Map k v) v

Traverse the value at the minimum key in a Map.

The key of the minimum element is available as the index of the traversal.

traverseAtMax :: SimpleIndexedTraversal k (Map k v) v

Traverse the value at the maximum key in a Map.

The key of the maximum element is available as the index of the traversal.