Data.Typeable.Lens

Description

Documentation

_cast :: (Typeable s, Typeable a) => Traversal' s a

A Traversal' for working with a cast of a Typeable value.

_gcast :: (Typeable s, Typeable a) => Traversal' (c s) (c a)

A Traversal' for working with a gcast of a Typeable value.