Data.Vector.Lens

Description

This module provides lenses and traversals for working with generic vectors.

Documentation

toVectorOf :: Getting (Endo [a]) s t a b -> s -> Vector a

Similar to toListOf, but returning a Vector.

Isomorphisms

vector :: Iso [a] [b] (Vector a) (Vector b)

Convert a list to a Vector (or back)

reversed :: Iso (Vector a) (Vector b) (Vector a) (Vector b)

Convert a Vector to a version with all the elements in the reverse order

forced :: Iso (Vector a) (Vector b) (Vector a) (Vector b)

Convert a Vector to a version that doesn't retain any extra memory.

Lenses

_head :: SimpleTraversal (Vector a) a

A Traversal reading and writing to the head of a Vector

>>> Vector.fromList [1,2,3] ^? _head
Just 1

>>> Vector.empty ^? _head
Nothing

>>> Vector.fromList "abc" & _head .~ 'Q'
fromList "Qbc"

_tail :: SimpleTraversal (Vector a) (Vector a)

A Traversal reading and writing to the tail of a Vector

>>> _tail .~ Vector.fromList [3,4,5] $ Vector.fromList [1,2]
fromList [1,3,4,5]

_last :: SimpleTraversal (Vector a) a

A Traversal reading and writing to the last element of a Vector

>>> Vector.fromList "abcde" ^? _last
Just 'e'

>>> Vector.empty ^? _last
Nothing

>>> Vector.fromList "abcde" & _last .~ 'Q'
fromList "abcdQ"

_init :: SimpleTraversal (Vector a) (Vector a)

A Traversal reading and replacing all but the a last element of a Vector

>>> Vector.fromList [1,2,3,4] ^? _init
Just (fromList [1,2,3])

>>> Vector.fromList "abcdef" & _init.mapped %~ succ
fromList "bcdeff"

sliced

Arguments

:: Int
i starting index

-> Int
n length

-> SimpleLens (Vector a) (Vector a)

sliced i n provides a lens that edits the n elements starting at index i from a lens.

This is only a valid lens if you do not change the length of the resulting Vector.

Attempting to return a longer or shorter vector will result in violations of the Lens laws.

Traversal of individual indices

ordinals :: [Int] -> SimpleIndexedTraversal Int (Vector a) a

This Traversal will ignore any duplicates in the supplied list of indices.

>>> toListOf (ordinals [1,3,2,5,9,10]) $ Vector.fromList [2,4..40]
[4,8,6,12,20,22]